Estimation function

Main estimation function for the tempered stable subordinator distribution (TSS), the classical tempered stable distribution (CTS), and the normal tempered stable distribution (NTS). It allows the user to select the preferred estimation method and several related options.

TemperedEstim(
  TemperedType = c("Classic", "Subordinator", "Normal"),
  EstimMethod = c("ML", "GMM", "Cgmm", "GMC"),
  data,
  theta0 = NULL,
  ComputeCov = FALSE,
  HandleError = TRUE,
  eps = 1e-06,
  algo = NULL,
  regularization = NULL,
  WeightingMatrix = NULL,
  t_scheme = NULL,
  alphaReg = NULL,
  t_free = NULL,
  subdivisions = NULL,
  IntegrationMethod = NULL,
  randomIntegrationLaw = NULL,
  s_min = NULL,
  s_max = NULL,
  ncond = NULL,
  IterationControl = NULL,
  ...
)

Arguments

TemperedType

A String. Either "Classic", "Subordinator", or "Normal"

EstimMethod

A String. Either "ML", "GMM", "Cgmm", or "GMC".

data

Data used to perform the estimation: numeric vector of length n.

theta0

A vector of numeric values corresponding to the pattern of the TemperedType.

ComputeCov

Logical flag: If set to TRUE, the asymptotic covariance matrix is computed. FALSE by default.

HandleError

Logical flag: If set to TRUE and if an error occurs during the estimation procedure, the computation will carry on and NA will be returned. Useful for Monte Carlo simulations.TRUE by default.

eps

Numerical error tolerance. 1e-06 by default.

algo

algorithm: For GMM: "2SGMM" is the two step GMM proposed by Hansen (1982). "CueGMM" and "ITGMM" are respectively the continuous updated and the iterative GMM proposed by Hansen, Eaton et Yaron (1996) and adapted to the continuum case. For GMC: "2SGMC", "CueGMC". For Cgmm: "2SCgmm", "CueCgmm", ....

regularization

regularization scheme to be used for moment methods, one of "Tikhonov" (Tikhonov), "LF" (Landweber-Fridmann) and "cut-off" (spectral cut-off).

WeightingMatrix

type of weighting matrix used to compute the objective function for the GMM and GMC methods, one of "OptAsym" (the optimal asymptotic), "DataVar" (the data driven, only for GMM) and "Id" (the identity matrix).

t_scheme

scheme used to select the points for the GMM method where the moment conditions are evaluated, one of "equally" (equally placed), "NonOptAr" (non optimal arithmetic placement), "uniformOpt" (uniform optimal placement), "ArithOpt" (arithmetic optimal placement), "Var Opt" (optimal variance placement) and "free" (users need to pass their own set of points in ...).

alphaReg

value of the regularisation parameter; numeric. Example Value could be ==0.01.

t_free

sequence, if t_scheme=="free".

subdivisions

Number of subdivisions used to compute the different integrals involved in the computation of the objective function for the Cgmm method (to minimise); numeric.

IntegrationMethod

Numerical integration method to be used to approximate the (vectorial) integrals for the Cgmm method. Users can choose between "Uniform" discretization or the "Simpson"'s rule (the 3-point Newton-Cotes quadrature rule).

randomIntegrationLaw

Probability measure associated to the Hilbert space spanned by the moment conditions for the Cgmm method.

s_min, s_max

Lower and Upper bounds of the interval where the moment conditions are considered for the Cgmm method; numeric.

ncond

Integer. Number of moment conditions (until order ncond) for the GMC method. Must not be less than 3 for TSS, 6 for CTS, 5 for NTS.

IterationControl

only used if algo = "IT..." or algo = "Cue..." to control the iterations. See Details.

...

Other arguments to be passed to the estimation function or the asymptotic confidence level.

Value

Object of a estim-class. See details for more information.

Details

TemperedType Detailed documentation of the individual tempered stable distributions can be viewed in the respective characteristic function. Use charTSS(), charCTS(), or charNTS().

Estimfct Additional parameters are needed for different estimation functions. These are listed below for each function. The list of additional parameters starts after the parameter eps in the parameter list.

For ML:

See usage of Maximum likelihood estimation in Kim et al. (2008). No additional parameters are needed.

For GMM:

Generalized Method of Moments by Feuerverger (1981). The parameters algo, alphaReg, regularization, WeightingMatrix, and t_scheme must be specified.

Parameter t_scheme: One of the most important features of this method is that it allows the user to choose how to place the points where the moment conditions are evaluated. One can choose among 6 different options. Depending on the option, further parameters have to be passed.

"equally":

equally placed points in min_t,max_t. When provided, user's min_t and max_t will be used (when Coinstrained == FALSE).

"NonOptAr":

non optimal arithmetic placement.

"uniformOpt":

uniform optimal placement.

"ArithOpt":

arithmetic optimal placement.

"Var Opt":

optimal variance placement as explained above.

"free":

user needs to pass own set of points in t_free.

Parameter WeightingMatrix: One can choose among 3 different options:

"OptAsym":

the optimal asymptotic choice.

"DataVar":

the covariance matrix of the data provided.

"Id":

the identity matrix.

For Cgmm:

Continuum Generalized Methods of Moments by Carrasco & Kotchoni (2017). The parameters algo, alphaReg, subdivisions, IntegrationMethod, randomIntegrationLaw, s_min, and s_max must be specified.

For GMC:

Generalized Method of Cumulants (GMC) by Massing, T. (2022). The parameters algo, alphaReg, regularization, WeightingMatrix, and ncond must be specified.

Estim-Class Class storing all the information about the estimation method; output of this function.

Slots of the return class

par:

Object of class "numeric"; Value of the estimated parameters.

par0:

Object of class "numeric"; Initial guess for the parameters.

vcov:

Object of class "matrix" representing the covariance matrix.

confint:

Object of class "matrix" representing the confidence interval computed at a specific level (attribute of the object).

data:

Object of class "numeric" used to compute the estimation.

sampleSize:

Object of class "numeric" ; length of the data.

others:

Object of class "list" ; more information about the estimation method.

duration:

Object of class "numeric" ; duration in seconds.

failure:

Object of class "numeric" representing the status of the procedure: 0 failure or 1 success.

method:

Object of class "character" description of the parameter used in the estimation.

IterationControl If algo = "IT..." or algo = "Cue..." the user can control each iteration by setting up the list IterationControl which contains the following elements:

NbIter

maximum number of iteration. The loop stops when NBIter is reached; default = 10.

PrintIterlogical

if set to TRUE, the value of the current parameter estimation is printed to the screen at each iteration; default = TRUE.

RelativeErrMax

the loop stops if the relative error between two consecutive estimation steps is smaller than RelativeErrMax; default = 1e-3.

Since this package is structurally based on the "StableEstim" package by Tarak Kharrat and Georgi N. Boshnakov, more detailed documentation can be found in their documentation.

References

Massing, T. (2023), 'Parametric Estimation of Tempered Stable Laws'

Kim, Y. s., Rachev, S. T., Bianchi, M. L. & Fabozzi, F. J. (2008), 'Financial market models with lévy processes and time-varying volatility' doi:10.1016/j.jbankfin.2007.11.004

Hansen, L. P. (1982), 'Large sample properties of generalized method of moments estimators' doi:10.2307/1912775

Hansen, L. P.; Heaton, J. & Yaron, A. (1996), 'Finite-Sample Properties of Some Alternative GMM Estimators' doi:10.1080/07350015.1996.10524656

Carrasco, M. & Kotchoni, R. (2017), 'Efficient estimation using the characteristic function' doi:10.1017/S0266466616000025

Kuechler, U. & Tappe, S. (2013), 'Tempered stable distribution and processes' doi:10.1016/j.spa.2013.06.012

Feuerverger, A. & McDunnough, P. (1981), 'On the efficiency of empirical characteristic function procedures' doi:10.1111/j.2517-6161.1981.tb01143.x

See also

Examples

# \donttest{
TemperedEstim(TemperedType = "Classic", EstimMethod = "ML",
               data = rCTS(2,1.5,1,1,1,1,0),
               theta0 = c(1.5,1,1,1,1,0) - 0.1);
#> *** Tempered Estim Classic, method Show *** 
#> ** Method ** 
#> [1] "ML_OptimAlgo=L-BFGS-B"
#> ** Parameters Estimation ** 
#>       alpha     delta +     delta -    lambda +    lambda -          mu 
#>  1.96360589  0.13631966  0.05299263  2.30419837  2.51008491 -0.61821795 
#> ** Covariance Matrix Estimation ** 
#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,]    0    0    0    0    0    0
#> [2,]    0    0    0    0    0    0
#> [3,]    0    0    0    0    0    0
#> [4,]    0    0    0    0    0    0
#> [5,]    0    0    0    0    0    0
#> [6,]    0    0    0    0    0    0
#> ** Confidence interval Estimation ** 
#> [1] "Confidence level= 0"
#> [1] "data length= 2"
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    0
#> [3,]    0    0
#> [4,]    0    0
#> [5,]    0    0
#> [6,]    0    0
#> attr(,"level")
#> [1] 0
#> ** Estimation time ** 
#> [1] " :duration= 0  h, 0  min, 4  sec. "
#> ** Estimation status ** 
#> success
#>  ******* End Show (Tempered Estim Classic) ******* 
TemperedEstim("Subordinator", "GMM", rTSS(20,0.5,1,1), algo = "2SGMM",
              alphaReg = 0.01, regularization = "cut-off",
              WeightingMatrix = "OptAsym", t_scheme = "free",
              t_free = seq(0.1,2,length.out = 12));
#> *** Tempered Estim Sub, method Show *** 
#> ** Method ** 
#> [1] "2SGMM_nb_t=12_alphaReg=0.01_regularization=cut-off_WeightingMatrix=OptAsym_t_scheme=free_OptimAlgo=nlminb"
#> ** Parameters Estimation ** 
#>     alpha     delta    lambda 
#> 0.8548119 0.2360579 0.6056726 
#> ** Covariance Matrix Estimation ** 
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]    0    0    0
#> [3,]    0    0    0
#> ** Confidence interval Estimation ** 
#> [1] "Confidence level= 0"
#> [1] "data length= 20"
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    0
#> [3,]    0    0
#> attr(,"level")
#> [1] 0
#> ** Estimation time ** 
#> [1] " :duration= 0  h, 0  min, 1  sec. "
#> ** Estimation status ** 
#> success
#>  ******* End Show (Tempered Estim Sub) ******* 
TemperedEstim("Normal", "Cgmm", rNTS(20,0.5,1,1,1,0), algo = "2SCgmm",
              alphaReg = 0.01, subdivisions = 50,
              IntegrationMethod = "Uniform", randomIntegrationLaw = "unif",
              s_min = 0, s_max= 1);
#> Warning: error during factorisation of matrix (dgefa);         singular matrix
#> Warning: steady-state not reached
#> *** Tempered Estim Normal, method Show *** 
#> ** Method ** 
#> [1] "Cgmm_algo=2SCgmm_alphaReg=0.01_OptimAlgo=nlminb_sd=50_IM=Uniform_RIL=unif_s_min=0_s_max=1"
#> ** Parameters Estimation ** 
#>        alpha         beta        delta       lambda           mu 
#> 0.9999675980 0.1030029195 0.0000521865 1.9169694997 1.3240533881 
#> ** Covariance Matrix Estimation ** 
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    0    0    0
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> ** Confidence interval Estimation ** 
#> [1] "Confidence level= 0"
#> [1] "data length= 20"
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    0
#> [3,]    0    0
#> [4,]    0    0
#> [5,]    0    0
#> attr(,"level")
#> [1] 0
#> ** Estimation time ** 
#> [1] " :duration= 0  h, 0  min, 9  sec. "
#> ** Estimation status ** 
#> success
#>  ******* End Show (Tempered Estim Normal) ******* 
TemperedEstim("Subordinator", "GMC", rTSS(20, 0.5, 1, 1), algo = "2SGMC",
              alphaReg = 0.01, WeightingMatrix = "OptAsym",
              regularization = "cut-off", ncond = 8, theta0 = c(0.5,1,1));
#> *** Tempered Estim Sub, method Show *** 
#> ** Method ** 
#> [1] "2SGMC_ncond=8_alphaReg=0.01_regularization=cut-off_WeightingMatrix=OptAsym_OptimAlgo=nlminb"
#> ** Parameters Estimation ** 
#>    alpha    delta   lambda 
#> 0.000001 3.816307 2.177499 
#> ** Covariance Matrix Estimation ** 
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]    0    0    0
#> [3,]    0    0    0
#> ** Confidence interval Estimation ** 
#> [1] "Confidence level= 0"
#> [1] "data length= 20"
#>      [,1] [,2]
#> [1,]    0    0
#> [2,]    0    0
#> [3,]    0    0
#> attr(,"level")
#> [1] 0
#> ** Estimation time ** 
#> [1] " :duration= 0  h, 0  min, 0  sec. "
#> ** Estimation status ** 
#> success
#>  ******* End Show (Tempered Estim Sub) ******* 
# }