Monte Carlo Simulation

Runs Monte Carlo simulation for a selected estimation method. The function can save results in a file.

TemperedEstim_Simulation(
  ParameterMatrix,
  SampleSizes = c(200, 1600),
  MCparam = 100,
  TemperedType = c("Classic", "Subordinator", "Normal"),
  Estimfct = c("ML", "GMM", "Cgmm", "GMC"),
  HandleError = TRUE,
  saveOutput = FALSE,
  SeedOptions = NULL,
  eps = 1e-06,
  algo = NULL,
  regularization = NULL,
  WeightingMatrix = NULL,
  t_scheme = NULL,
  alphaReg = NULL,
  t_free = NULL,
  subdivisions = NULL,
  IntegrationMethod = NULL,
  randomIntegrationLaw = NULL,
  s_min = NULL,
  s_max = NULL,
  ncond = NULL,
  IterationControl = NULL,
  methodR = "AR",
  ...
)

Arguments

ParameterMatrix

The matrix is to be composed of vectors, row by row. Each vector must fit the pattern of theta of the TemperedType.

SampleSizes

Sample sizes to be used to simulate the data. By default, we use 200 (small sample size) and 1600 (large sample size); vector of integer.

MCparam

Number of Monte Carlo simulation for each couple of parameter, default=100; integer

TemperedType

A String. Either "Classic", "Subordinator", or "Normal".

Estimfct

The estimation function to be used. A String. Either "ML", "GMM", "Cgmm", or "GMC".

HandleError

Logical flag: if set to TRUE, the simulation doesn't stop when an error in the estimation function is encountered. A vector of (size 4) NA is saved and the the simulation carries on. See details.

saveOutput

Logical flag: if set to TRUE, a csv file (for each couple of parameter) with the the estimation information is saved in the current directory. See details.

SeedOptions

List to control the seed generation. See details.

eps

Numerical error tolerance. 1e-06 by default.

algo

algorithm: For GMM: "2SGMM" is the two step GMM proposed by Hansen (1982). "CueGMM" and "ITGMM" are respectively the continuous updated and the iterative GMM proposed by Hansen, Eaton et Yaron (1996) and adapted to the continuum case. For GMC: "2SGMC", "CueGMC". For Cgmm: "2SCgmm", "CueCgmm", ....

regularization

regularization scheme to be used for moment methods, one of "Tikhonov" (Tikhonov), "LF" (Landweber-Fridmann) and "cut-off" (spectral cut-off).

WeightingMatrix

type of weighting matrix used to compute the objective function for the GMM and GMC methods, one of "OptAsym" (the optimal asymptotic), "DataVar" (the data driven, only for GMM) and "Id" (the identity matrix).

t_scheme

scheme used to select the points for the GMM method where the moment conditions are evaluated, one of "equally" (equally placed), "NonOptAr" (non optimal arithmetic placement), "uniformOpt" (uniform optimal placement), "ArithOpt" (arithmetic optimal placement), "Var Opt" (optimal variance placement) and "free" (users need to pass their own set of points in ...).

alphaReg

value of the regularisation parameter; numeric. Example Value could be ==0.01.

t_free

sequence, if t_scheme=="free".

subdivisions

Number of subdivisions used to compute the different integrals involved in the computation of the objective function for the Cgmm method (to minimise); numeric.

IntegrationMethod

Numerical integration method to be used to approximate the (vectorial) integrals for the Cgmm method. Users can choose between "Uniform" discretization or the "Simpson"'s rule (the 3-point Newton-Cotes quadrature rule).

randomIntegrationLaw

Probability measure associated to the Hilbert space spanned by the moment conditions for the Cgmm method.

s_min, s_max

Lower and Upper bounds of the interval where the moment conditions are considered for the Cgmm method; numeric.

ncond

Integer. Number of moment conditions (until order ncond) for the GMC method. Must not be less than 3 for TSS, 6 for CTS, 5 for NTS.

IterationControl

only used if algo = "IT..." or algo = "Cue..." to control the iterations. See Details.

methodR

A string. It selects the method in rCTS/rNTS/rTTS. "AR" by default.

...

Other arguments to be passed to the estimation function.

Value

If saveOutput == FALSE, the return object is a list of 2. Results of the simulation are listed in $outputMat. If saveOutput == TRUE, only a csv file is saved and nothing is returned.

Details

Error Handling It is advisable to set it to TRUE when user is planning to launch long simulations as it will prevent the procedure to stop if an error occurs for one sample data. The estimation function will produce a vector of NA as estimated parameters related to this (error generating) sample data and move on to the next Monte Carlo step.

Output file Setting saveOutput to TRUE will have the side effect of saving a csv file in the working directory. This file will have MCparam*length(SampleSizes) lines and its columns will be:

alphaT, ...:

the true value of the parameters.

data size:

the sample size used to generate the simulated data.

seed:

the seed value used to generate the simulated data.

alphaE, ...:

the estimate of the parameters.

failure:

binary: 0 for success, 1 for failure.

time:

estimation running time in seconds.

The file name is informative to let the user identify the value of the true parameters, the MC parameters as well as the options selected for the estimation method. The csv file is updated after each MC estimation which is useful when the simulation stops before it finishes.

SeedOptions If users does not want to control the seed generation, they could ignore this argument (default value NULL). This argument can be more useful when they wants to cut the simulation (even for one parameter value) into pieces. In that case, they can control which part of the seed vector they want to use.

MCtot:

total values of MC simulations in the entire process.

seedStart:

starting index in the seed vector. The vector extracted will be of size MCparam.

Estimfct Additional parameters are needed for different estimation functions. These are listed below for each function. The list of additional parameters starts after the parameter eps in the parameter list.

For ML:

See usage of Maximum likelihood estimation in Kim et al. (2008).No additional parameters are needed.

For GMM:

Generalized Method of Moments by Feuerverger (1981). The parameters algo, alphaReg, regularization, WeightingMatrix, and t_scheme must be specified.

Parameter t_scheme: One of the most important features of this method is that it allows the user to choose how to place the points where the moment conditions are evaluated. One can choose among 6 different options. Depending on the option, further parameters have to be passed.

"equally":

equally placed points in min_t,max_t. When provided, user's min_t and max_t will be used (when Coinstrained == FALSE).

"NonOptAr":

non optimal arithmetic placement.

"uniformOpt":

uniform optimal placement.

"ArithOpt":

arithmetic optimal placement.

"Var Opt":

optimal variance placement as explained above.

"free":

user needs to pass own set of points in t_free.

Parameter WeightingMatrix: One can choose among 3 different options:

"OptAsym":

the optimal asymptotic choice.

"DataVar":

the covariance matrix of the data provided.

"Id":

the identity matrix.

For Cgmm:

Continuum Generalized Methods of Moments by Carrasco & Kotchoni (2017). The parameters algo, alphaReg, subdivisions, IntegrationMethod, randomIntegrationLaw, s_min, and s_max must be specified.

For GMC:

Generalized Method of Cumulants (GMC) by Massing, T. (2022). The parameters algo, alphaReg, regularization, WeightingMatrix, and ncond must be specified.

IterationControl If algo = "IT..." or algo = "Cue..." the user can control each iteration by setting up the list IterationControl which contains the following elements:

NbIter

maximum number of iteration. The loop stops when NBIter is reached; default = 10.

PrintIterlogical

if set to TRUE, the value of the current parameter estimation is printed to the screen at each iteration; default = TRUE.

RelativeErrMax

the loop stops if the relative error between two consecutive estimation steps is smaller than RelativeErrMax; default = 1e-3.

Parallelization Parallelization of the function is possible with using parallelizeMCsimulation(). If someone wants to parallelize the function manually, the parameter MCparam must be set to 1 and the parameter SeedOption must be changed for each iteration.

Since this package is structurally based on the "StableEstim" package by Tarak Kharrat and Georgi N. Boshnakov, more detailed documentation can be found in their documentation.

References

Massing, T. (2022), 'Parametric Estimation of Tempered Stable Laws'

Kim, Y. s.; Rachev, S. T.; Bianchi, M. L. & Fabozzi, F. J. (2008), 'Financial market models with lévy processes and time-varying volatility' doi:10.1016/j.jbankfin.2007.11.004

Hansen, L. P. (1982), 'Large sample properties of generalized method of moments estimators' doi:10.2307/1912775

Hansen, L. P.; Heaton, J. & Yaron, A. (1996), 'Finite-Sample Properties of Some Alternative GMM Estimators' doi:10.1080/07350015.1996.10524656

Feuerverger, A. & McDunnough, P. (1981), 'On the efficiency of empirical characteristic function procedures' doi:10.1111/j.2517-6161.1981.tb01143.x

Carrasco, M. & Kotchoni, R. (2017), 'Efficient estimation using the characteristic function' doi:10.1017/S0266466616000025 ;

Kuechler, U. & Tappe, S. (2013), 'Tempered stable distribution and processes' doi:10.1016/j.spa.2013.06.012

See also

Examples

# \donttest{
TemperedEstim_Simulation(ParameterMatrix = rbind(c(1.5,1,1,1,1,0),
                                                 c(0.5,1,1,1,1,0)),
                         SampleSizes = c(10), MCparam = 10,
                         TemperedType = "Classic", Estimfct = "ML",
                         saveOutput = FALSE)
#> ---------------- Alpha=1.5 *** DeltaP=1 *** DeltaM=1 *** LambdaP=1 *** LambdaM=1 *** mu=0 --------------- 
#> *** Iter 1/10 *** Estimated Remaining Time: 0h3min10sec. *** 
#> *** Iter 2/10 *** Estimated Remaining Time: 0h3min12sec. *** 
#> *** Iter 3/10 *** Estimated Remaining Time: 0h6min29sec. *** 
#> *** Iter 4/10 *** Estimated Remaining Time: 0h4min46sec. *** 
#> Warning: NaNs produced
#> Warning: NaNs produced
#> *** Iter 5/10 *** Estimated Remaining Time: 0h1min13sec. *** 
#> Warning: NaNs produced
#> *** Iter 6/10 *** Estimated Remaining Time: 0h0min46sec. *** 
#> *** Iter 7/10 *** Estimated Remaining Time: 0h1min20sec. *** 
#> Warning: NaNs produced
#> *** Iter 8/10 *** Estimated Remaining Time: 0h0min23sec. *** 
#> Warning: NaNs produced
#> *** Iter 9/10 *** Estimated Remaining Time: 0h0min12sec. *** 
#> Warning: NaNs produced
#> *** Iter 10/10 *** Estimated Remaining Time: 0h0min0sec. *** 
#> ---------------- Alpha=0.5 *** DeltaP=1 *** DeltaM=1 *** LambdaP=1 *** LambdaM=1 *** mu=0 --------------- 
#> *** Iter 1/10 *** Estimated Remaining Time: 0h6min4sec. *** 
#> Warning: NaNs produced
#> *** Iter 2/10 *** Estimated Remaining Time: 0h0min55sec. *** 
#> Warning: NaNs produced
#> *** Iter 3/10 *** Estimated Remaining Time: 0h0min20sec. *** 
#> Warning: NaNs produced
#> *** Iter 4/10 *** Estimated Remaining Time: 0h1min17sec. *** 
#> Warning: NaNs produced
#> *** Iter 5/10 *** Estimated Remaining Time: 0h0min16sec. *** 
#> *** Iter 6/10 *** Estimated Remaining Time: 0h2min44sec. *** 
#> Warning: NaNs produced
#> *** Iter 7/10 *** Estimated Remaining Time: 0h0min52sec. *** 
#> *** Iter 8/10 *** Estimated Remaining Time: 0h0min50sec. *** 
#> *** Iter 9/10 *** Estimated Remaining Time: 0h0min35sec. *** 
#> Warning: NaNs produced
#> *** Iter 10/10 *** Estimated Remaining Time: 0h0min0sec. *** 
#> $outputMat
#>       alphaT delta+T delta-T lambda+T lambda-T muT data size seed    alphaE
#>  [1,]    1.5       1       1        1        1   0        10   29 0.0000010
#>  [2,]    1.5       1       1        1        1   0        10   23 0.0000010
#>  [3,]    1.5       1       1        1        1   0        10   19 0.0000010
#>  [4,]    1.5       1       1        1        1   0        10   21 0.9785304
#>  [5,]    1.5       1       1        1        1   0        10   18       NaN
#>  [6,]    1.5       1       1        1        1   0        10   20       NaN
#>  [7,]    1.5       1       1        1        1   0        10    8 0.0000010
#>  [8,]    1.5       1       1        1        1   0        10    1       NaN
#>  [9,]    1.5       1       1        1        1   0        10    9       NaN
#> [10,]    1.5       1       1        1        1   0        10    2       NaN
#> [11,]    0.5       1       1        1        1   0        10   29 0.0000010
#> [12,]    0.5       1       1        1        1   0        10   23       NaN
#> [13,]    0.5       1       1        1        1   0        10   19       NaN
#> [14,]    0.5       1       1        1        1   0        10   21       NaN
#> [15,]    0.5       1       1        1        1   0        10   18       NaN
#> [16,]    0.5       1       1        1        1   0        10   20 0.0000010
#> [17,]    0.5       1       1        1        1   0        10    8       NaN
#> [18,]    0.5       1       1        1        1   0        10    1 0.0000010
#> [19,]    0.5       1       1        1        1   0        10    9 0.0000010
#> [20,]    0.5       1       1        1        1   0        10    2       NaN
#>          delta+E    delta-E    lambda+E  lambda-E         muE failure  time
#>  [1,]  0.6422503  0.4339281 1.113048701 0.3424624 -0.51469069       0 21.11
#>  [2,]  0.9534512  0.1874224 0.820659245 0.3573659  0.98227225       0 24.03
#>  [3,]  0.0000010 16.2718628 6.507556463 2.6364859  1.11577311       0 55.56
#>  [4,]  0.0000010  6.0584029 0.019151849 3.5772032 -0.07073078       0 47.60
#>  [5,]        NaN        NaN         NaN       NaN         NaN       1  0.00
#>  [6,]        NaN        NaN         NaN       NaN         NaN       1  0.00
#>  [7,] 16.7686506  0.0000010 2.822096764 5.8613166  0.53310815       0 26.64
#>  [8,]        NaN        NaN         NaN       NaN         NaN       1  0.00
#>  [9,]        NaN        NaN         NaN       NaN         NaN       1  0.00
#> [10,]        NaN        NaN         NaN       NaN         NaN       1  0.00
#> [11,]  0.3294759  0.7038471 1.168866691 2.7851732 -0.09727607       0 40.46
#> [12,]        NaN        NaN         NaN       NaN         NaN       1  0.00
#> [13,]        NaN        NaN         NaN       NaN         NaN       1  0.00
#> [14,]        NaN        NaN         NaN       NaN         NaN       1  0.00
#> [15,]        NaN        NaN         NaN       NaN         NaN       1  0.00
#> [16,]  0.0000010 15.3240499 0.002306115 3.1896860 -0.20533131       0 41.06
#> [17,]        NaN        NaN         NaN       NaN         NaN       1  0.00
#> [18,]  1.0845178  0.0000010 0.953728035 1.6255790  0.09580610       0 24.86
#> [19,]  0.2427847  0.6376896 0.753585927 0.6469205  0.28425570       0 34.83
#> [20,]        NaN        NaN         NaN       NaN         NaN       1  0.00
#> 

TemperedEstim_Simulation(ParameterMatrix = rbind(c(1.5,1,1,1,1,0)),
                         SampleSizes = c(40), MCparam = 40,
                         TemperedType = "Classic", Estimfct = "GMM",
                         saveOutput = FALSE, algo = "2SGMM",
                         regularization = "cut-off",
                         WeightingMatrix = "OptAsym", t_scheme = "free",
                         alphaReg = 0.01,
                         t_free = seq(0.1,2,length.out=12))
#> ---------------- Alpha=1.5 *** DeltaP=1 *** DeltaM=1 *** LambdaP=1 *** LambdaM=1 *** mu=0 --------------- 
#> *** Iter 1/40 *** Estimated Remaining Time: 0h1min32sec. *** 
#> *** Iter 2/40 *** Estimated Remaining Time: 0h1min53sec. *** 
#> *** Iter 3/40 *** Estimated Remaining Time: 0h1min4sec. *** 
#> *** Iter 4/40 *** Estimated Remaining Time: 0h0min60sec. *** 
#> *** Iter 5/40 *** Estimated Remaining Time: 0h0min33sec. *** 
#> *** Iter 6/40 *** Estimated Remaining Time: 0h2min15sec. *** 
#> *** Iter 7/40 *** Estimated Remaining Time: 0h0min36sec. *** 
#> *** Iter 8/40 *** Estimated Remaining Time: 0h1min53sec. *** 
#> *** Iter 9/40 *** Estimated Remaining Time: 0h1min38sec. *** 
#> *** Iter 10/40 *** Estimated Remaining Time: 0h0min31sec. *** 
#> *** Iter 11/40 *** Estimated Remaining Time: 0h0min37sec. *** 
#> *** Iter 12/40 *** Estimated Remaining Time: 0h0min32sec. *** 
#> *** Iter 13/40 *** Estimated Remaining Time: 0h0min27sec. *** 
#> *** Iter 14/40 *** Estimated Remaining Time: 0h0min21sec. *** 
#> *** Iter 15/40 *** Estimated Remaining Time: 0h0min38sec. *** 
#> *** Iter 16/40 *** Estimated Remaining Time: 0h0min30sec. *** 
#> *** Iter 17/40 *** Estimated Remaining Time: 0h0min17sec. *** 
#> *** Iter 18/40 *** Estimated Remaining Time: 0h0min21sec. *** 
#> *** Iter 19/40 *** Estimated Remaining Time: 0h0min45sec. *** 
#> *** Iter 20/40 *** Estimated Remaining Time: 0h0min25sec. *** 
#> *** Iter 21/40 *** Estimated Remaining Time: 0h0min21sec. *** 
#> *** Iter 22/40 *** Estimated Remaining Time: 0h0min17sec. *** 
#> *** Iter 23/40 *** Estimated Remaining Time: 0h0min54sec. *** 
#> *** Iter 24/40 *** Estimated Remaining Time: 0h0min26sec. *** 
#> *** Iter 25/40 *** Estimated Remaining Time: 0h0min16sec. *** 
#> *** Iter 26/40 *** Estimated Remaining Time: 0h0min20sec. *** 
#> *** Iter 27/40 *** Estimated Remaining Time: 0h0min32sec. *** 
#> *** Iter 28/40 *** Estimated Remaining Time: 0h0min12sec. *** 
#> *** Iter 29/40 *** Estimated Remaining Time: 0h0min21sec. *** 
#> *** Iter 30/40 *** Estimated Remaining Time: 0h0min8sec. *** 
#> *** Iter 31/40 *** Estimated Remaining Time: 0h0min8sec. *** 
#> *** Iter 32/40 *** Estimated Remaining Time: 0h0min10sec. *** 
#> *** Iter 33/40 *** Estimated Remaining Time: 0h0min18sec. *** 
#> *** Iter 34/40 *** Estimated Remaining Time: 0h0min9sec. *** 
#> *** Iter 35/40 *** Estimated Remaining Time: 0h0min7sec. *** 
#> *** Iter 36/40 *** Estimated Remaining Time: 0h0min4sec. *** 
#> *** Iter 37/40 *** Estimated Remaining Time: 0h0min5sec. *** 
#> *** Iter 38/40 *** Estimated Remaining Time: 0h0min3sec. *** 
#> *** Iter 39/40 *** Estimated Remaining Time: 0h0min2sec. *** 
#> *** Iter 40/40 *** Estimated Remaining Time: 0h0min0sec. *** 
#> $outputMat
#>       alphaT delta+T delta-T lambda+T lambda-T muT data size seed       alphaE
#>  [1,]    1.5       1       1        1        1   0        40   93 1.349162e-06
#>  [2,]    1.5       1       1        1        1   0        40   87 1.558819e-06
#>  [3,]    1.5       1       1        1        1   0        40   83 1.000000e-06
#>  [4,]    1.5       1       1        1        1   0        40   95 1.540780e+00
#>  [5,]    1.5       1       1        1        1   0        40   21 1.016411e-06
#>  [6,]    1.5       1       1        1        1   0        40   50 1.717760e+00
#>  [7,]    1.5       1       1        1        1   0        40   60 1.000000e-06
#>  [8,]    1.5       1       1        1        1   0        40   84 1.091234e+00
#>  [9,]    1.5       1       1        1        1   0        40   28 9.038206e-01
#> [10,]    1.5       1       1        1        1   0        40  104 1.000000e-06
#> [11,]    1.5       1       1        1        1   0        40   97 1.000562e-06
#> [12,]    1.5       1       1        1        1   0        40  105 1.000000e-06
#> [13,]    1.5       1       1        1        1   0        40   66 1.000000e-06
#> [14,]    1.5       1       1        1        1   0        40   89 1.000000e-06
#> [15,]    1.5       1       1        1        1   0        40   90 1.000000e-06
#> [16,]    1.5       1       1        1        1   0        40   19 1.000000e-06
#> [17,]    1.5       1       1        1        1   0        40   37 1.000000e-06
#> [18,]    1.5       1       1        1        1   0        40    3 1.000000e-06
#> [19,]    1.5       1       1        1        1   0        40   69 1.000700e-06
#> [20,]    1.5       1       1        1        1   0        40  109 1.016337e+00
#> [21,]    1.5       1       1        1        1   0        40   45 1.000000e-06
#> [22,]    1.5       1       1        1        1   0        40   47 1.005792e-06
#> [23,]    1.5       1       1        1        1   0        40   61 1.000000e-06
#> [24,]    1.5       1       1        1        1   0        40   33 1.000000e-06
#> [25,]    1.5       1       1        1        1   0        40  120 1.000000e-06
#> [26,]    1.5       1       1        1        1   0        40   32 1.000000e-06
#> [27,]    1.5       1       1        1        1   0        40   16 2.237671e-06
#> [28,]    1.5       1       1        1        1   0        40   40 1.000000e-06
#> [29,]    1.5       1       1        1        1   0        40   94 1.009022e-06
#> [30,]    1.5       1       1        1        1   0        40    7 1.000000e-06
#> [31,]    1.5       1       1        1        1   0        40   96 1.000000e-06
#> [32,]    1.5       1       1        1        1   0        40   27 1.028639e-06
#> [33,]    1.5       1       1        1        1   0        40  103 1.000000e-06
#> [34,]    1.5       1       1        1        1   0        40   75 1.000173e-06
#> [35,]    1.5       1       1        1        1   0        40   91 1.790607e-06
#> [36,]    1.5       1       1        1        1   0        40  108 1.000015e-06
#> [37,]    1.5       1       1        1        1   0        40   26 1.000000e-06
#> [38,]    1.5       1       1        1        1   0        40    2 1.001066e-06
#> [39,]    1.5       1       1        1        1   0        40   18 1.000178e-06
#> [40,]    1.5       1       1        1        1   0        40  117 8.671253e-01
#>           delta+E      delta-E     lambda+E     lambda-E           muE failure
#>  [1,]  0.00000100 63.725821799 2.241021e+01 5.489916e+00  0.1084625343       0
#>  [2,]  0.00000100 77.154164788 1.382010e+02 5.161094e+00 -0.1099511565       0
#>  [3,] 15.67640225  0.000001000 2.053828e+00 7.002056e+00  0.3003841274       0
#>  [4,]  0.94526358  0.000001000 2.573962e-01 4.259130e+01  0.4047807050       0
#>  [5,]  0.00000100  5.244718810 3.464605e+00 1.171329e+00 -0.0701746820       0
#>  [6,]  2.55902335  0.431218969 4.892124e+02 7.112388e-02 -0.2191348321       0
#>  [7,]  0.09851492  3.767191860 2.385211e-01 1.318416e+00 -0.2606247214       0
#>  [8,]  2.03827930 29.749247918 5.192058e+01 2.089154e+01 -0.0003813403       0
#>  [9,]  1.35593913 52.641666115 5.830320e-01 2.199883e+01  0.1986158376       0
#> [10,]  1.35822235  3.019622572 1.063180e+00 1.184617e+00 -0.0893606852       0
#> [11,]  0.04760970  2.382716890 2.359113e-01 8.212715e-01 -0.0340304049       0
#> [12,]  1.64316270  0.375804442 8.822020e-01 5.400657e-01 -0.9470540175       0
#> [13,]  0.57772700  3.008501400 6.148035e-01 9.336172e-01 -0.5415068924       0
#> [14,]  1.61289428  0.883167011 1.037478e+00 7.574171e-01  0.0415862362       0
#> [15,]  2.55109122  1.010439827 9.469633e-01 1.812502e+00 -0.3739395214       0
#> [16,] 27.91364048 25.409128185 8.521636e+00 2.589438e+00  0.1509102497       0
#> [17,]  0.58923868  1.010288916 5.957892e-01 6.868326e-01 -0.3522906423       0
#> [18,]  1.24974748  0.709585776 8.314492e-01 4.301232e-01 -0.3729344304       0
#> [19,] 58.90734476  0.000001000 4.106500e+00 5.953210e+00 -0.2170581937       0
#> [20,]  0.17863996  2.741960941 1.835192e-05 1.651943e+00  1.1204750128       0
#> [21,]  0.45169100  1.877917428 5.201738e-01 7.320815e-01  0.4020238475       0
#> [22,]  2.85439085  0.605268821 1.555163e+00 5.579310e-01 -0.3603878592       0
#> [23,] 83.71974810  0.000001000 5.847856e+00 8.484978e+02 -0.6756456186       0
#> [24,]  0.00000100  9.896024498 6.885711e+00 1.571422e+00 -0.1497665469       0
#> [25,]  1.81381164  1.439255819 1.009767e+00 8.745029e-01 -0.4976871714       0
#> [26,] 19.01878086  0.000001000 2.467491e+00 2.553303e+01  0.5435369047       0
#> [27,] 43.60827464  0.001459519 3.914419e+00 9.597163e+01  0.3410550557       0
#> [28,]  0.88829565  0.694396387 5.203525e-01 6.752082e-01  0.6984349043       0
#> [29,]  0.00000100 31.031728594 8.351390e+00 3.467155e+00  0.1664807510       0
#> [30,]  0.33526497  4.946632779 4.768227e-01 2.189378e+00  0.0607051923       0
#> [31,]  0.00000100  3.450270875 2.900190e+00 9.320761e-01 -0.0760130928       0
#> [32,]  4.84194554  6.187495420 2.790182e+00 1.492717e+00  0.0423203719       0
#> [33,]  5.72523571  0.000001000 1.123323e+00 3.216625e+00  0.2345551672       0
#> [34,] 26.52818973  0.000001000 3.699984e+00 1.286241e+01  0.1840449276       0
#> [35,]  1.07996726  2.691774498 6.136334e-01 1.666735e+00  0.0744312847       0
#> [36,]  0.75219601  1.171113700 6.848029e-01 7.685355e-01 -0.0455350749       0
#> [37,]  0.37002252  4.054576635 4.961088e-01 1.094220e+00 -0.2204668825       0
#> [38,] 14.86926782  0.000001000 2.143761e+00 2.071889e+01  0.6204852734       0
#> [39,]  2.24223771 44.932094692 7.444241e-01 1.291089e+01 -0.1380219374       0
#> [40,] 21.57270568  0.000001000 4.549722e+00 7.570004e-03 -0.5587868265       0
#>       time
#>  [1,] 2.36
#>  [2,] 2.97
#>  [3,] 1.71
#>  [4,] 1.64
#>  [5,] 0.95
#>  [6,] 3.95
#>  [7,] 1.06
#>  [8,] 3.51
#>  [9,] 3.16
#> [10,] 1.05
#> [11,] 1.28
#> [12,] 1.14
#> [13,] 0.98
#> [14,] 0.82
#> [15,] 1.53
#> [16,] 1.22
#> [17,] 0.76
#> [18,] 0.97
#> [19,] 2.11
#> [20,] 1.25
#> [21,] 1.09
#> [22,] 0.94
#> [23,] 3.20
#> [24,] 1.63
#> [25,] 1.06
#> [26,] 1.44
#> [27,] 2.49
#> [28,] 0.97
#> [29,] 1.87
#> [30,] 0.85
#> [31,] 0.85
#> [32,] 1.27
#> [33,] 2.53
#> [34,] 1.52
#> [35,] 1.46
#> [36,] 1.00
#> [37,] 1.57
#> [38,] 1.73
#> [39,] 2.41
#> [40,] 1.81
#> 

TemperedEstim_Simulation(ParameterMatrix = rbind(c(1.45,0.55,1,1,1,0)),
                         SampleSizes = c(4), MCparam = 4,
                         TemperedType = "Classic", Estimfct = "Cgmm",
                         saveOutput = FALSE, algo = "2SCgmm",
                         alphaReg = 0.01, subdivisions = 50,
                         IntegrationMethod = "Uniform",
                         randomIntegrationLaw = "unif",
                         s_min = 0, s_max= 1)
#> ---------------- Alpha=1.45 *** DeltaP=0.55 *** DeltaM=1 *** LambdaP=1 *** LambdaM=1 *** mu=0 --------------- 
#> *** Iter 1/4 *** Estimated Remaining Time: 0h0min4sec. *** 
#> *** Iter 2/4 *** Estimated Remaining Time: 0h0min8sec. *** 
#> *** Iter 3/4 *** Estimated Remaining Time: 0h0min5sec. *** 
#> *** Iter 4/4 *** Estimated Remaining Time: 0h0min0sec. *** 
#> $outputMat
#>      alphaT delta+T delta-T lambda+T lambda-T muT data size seed   alphaE
#> [1,]   1.45    0.55       1        1        1   0         4    7 0.542067
#> [2,]   1.45    0.55       1        1        1   0         4    3 1.301437
#> [3,]   1.45    0.55       1        1        1   0         4    5 1.777274
#> [4,]   1.45    0.55       1        1        1   0         4    2 1.292597
#>        delta+E   delta-E lambda+E lambda-E        muE failure time
#> [1,] 0.1204020 0.3735694 1.021255 1.306307 -1.0112375       0 1.43
#> [2,] 0.4680792 0.0000010 1.433965 1.658598  0.7931090       0 3.90
#> [3,] 0.1641245 0.0000010 1.661967 1.858982 -0.7489969       0 4.61
#> [4,] 0.0000010 2.3544647 1.507313 3.251520 -0.9111893       0 8.12
#> 

TemperedEstim_Simulation(ParameterMatrix = rbind(c(1.45,0.55,1,1,1,0)),
                         SampleSizes = c(4), MCparam = 4,
                         TemperedType = "Classic", Estimfct = "GMC",
                         saveOutput = FALSE, algo = "2SGMC",
                         alphaReg = 0.01, WeightingMatrix = "OptAsym",
                         regularization = "cut-off", ncond = 8)
#> ---------------- Alpha=1.45 *** DeltaP=0.55 *** DeltaM=1 *** LambdaP=1 *** LambdaM=1 *** mu=0 --------------- 
#> *** Iter 1/4 *** Estimated Remaining Time: 0h0min4sec. *** 
#> *** Iter 2/4 *** Estimated Remaining Time: 0h0min5sec. *** 
#> *** Iter 3/4 *** Estimated Remaining Time: 0h0min1sec. *** 
#> *** Iter 4/4 *** Estimated Remaining Time: 0h0min0sec. *** 
#> $outputMat
#>      alphaT delta+T delta-T lambda+T lambda-T muT data size seed    alphaE
#> [1,]   1.45    0.55       1        1        1   0         4    7 1.0071494
#> [2,]   1.45    0.55       1        1        1   0         4    3 0.2554630
#> [3,]   1.45    0.55       1        1        1   0         4    5 0.9999964
#> [4,]   1.45    0.55       1        1        1   0         4    2 1.6603913
#>        delta+E      delta-E  lambda+E  lambda-E         muE failure time
#> [1,]  0.000001 1.000000e-06 11.093087 10.168951  -0.8131496       0 1.46
#> [2,] 13.182909 5.849044e+04  2.993486 93.064799 -11.5293848       0 2.41
#> [3,]  0.000001 1.000000e-06  6.911297  5.968896  -0.6495758       0 0.65
#> [4,]  0.000001 3.237970e-03  1.854717  2.373008  -0.4168512       0 0.39
#> 
# }