Cumulative probability function of the rapidly decreasing tempered stable (RDTS) distribution

The cumulative probability distribution function (CDF) of the rapidly decreasing tempered stable distribution.

pRDTS(
  q,
  alpha = NULL,
  delta = NULL,
  lambdap = NULL,
  lambdam = NULL,
  mu = NULL,
  theta = NULL,
  dens_method = "FFT",
  a = -130,
  b = 130,
  nf = 2048,
  ...
)

Arguments

q

A numeric vector of quantiles.

alpha

Stability parameter. A real number between 0 and 2.

delta

Scale parameter for the left tail. A real number > 0.

lambdap

Tempering parameter for the right tail. A real number > 0.

lambdam

Tempering parameter for the left tail. A real number > 0.

mu

A location parameter, any real number.

theta

Parameters stacked as a vector.

dens_method

Algorithm for numerical evaluation. Currently, only "FFT" available.

a

Starting point of FFT, if dens_method == "FFT". -20 by default.

b

Ending point of FFT, if dens_method == "FFT". 20 by default.

nf

Pieces the transformation is divided in. Limited to power-of-two size.

...

Possibility to modify stats::integrate().

Value

As q is a numeric vector, the return value is also a numeric vector of probabilities.

Details

theta denotes the parameter vector (alpha, delta, lambdap, lambdam, mu). Either provide the parameters individually OR provide theta. The function integrates the PDF numerically with integrate().

See also