Function to generate random variates of NTS distribution.

Generates n random numbers distributed according of the normal tempered stable distribution.

rNTS(
  n,
  alpha = NULL,
  beta = NULL,
  delta = NULL,
  lambda = NULL,
  mu = NULL,
  theta = NULL,
  method = "AR",
  k = 10000
)

Arguments

n

sample size (integer).

alpha

A real number between 0 and 1.

beta

A gap holder.

delta

A real number > 0.

lambda

A real number > 0.

mu

A location parameter, any real number.

theta

A vector of all other arguments.

method

A String. Either "AR" or "SR". "AR" by default.

k

integer: the number of replications, if method == "SR". 10000 by default.

Value

Generates n random numbers.

Details

theta denotes the parameter vector (alpha, beta, delta, lambda, mu). Either provide the parameters individually OR provide theta. Works by a normal variance-mean mixture with a TSS distribution. Method parameter is for the method of simulating the TSS random variable, see the rTSS() function. "AR" stands for the Acceptance-Rejection Method and "SR" for a truncated infinite shot noise series representation. "AR" is the standard method used.

For more details, see references.

References

Massing, Till (2022), 'Parametric Estimation of Tempered Stable Laws'

Kawai, R & Masuda, H (2011), 'On simulation of tempered stable random variates' doi:10.1016/j.cam.2010.12.014

See also

Examples

rNTS(100, 0.5, 1,1,1,1)
#>   [1]  0.3986615  0.8692852  0.4723544  1.6404555  1.7132874  3.9206922
#>   [7]  3.7312978  1.5743540  2.6904323  3.6668598  4.1336013  1.3603767
#>  [13]  1.7667871  4.2744329  1.3212757  2.1369741  2.3333982  2.0335539
#>  [19]  1.3558653  2.4100920  1.7804172  1.6233614  7.2258867  5.1220599
#>  [25]  5.8911413  2.9816986  3.1284117  2.7836234  3.7392594  2.5711610
#>  [31]  1.6455587  3.3295346  1.6502757  2.9053164  2.2212419  1.5337705
#>  [37]  2.5170115  1.1453093  4.0216954  3.8802222  1.7611299  1.9233605
#>  [43]  0.4218459  2.9950148  5.8968669  1.5952703 -0.7745171  2.2161983
#>  [49]  3.9344175  4.4946267  1.6351951  5.6000233  2.4421089  1.6348407
#>  [55]  2.5400751  4.8208260  3.2277660  4.9125952  0.8561953  1.8051706
#>  [61]  0.8285870  1.7916377  2.0014203  4.9340508  3.5227276  2.0047274
#>  [67]  1.5394965  2.5568700  3.0949304  1.4036772  2.8526902  3.5140480
#>  [73]  4.1839562  4.8593027  5.1886800  0.9044691  2.0535937  1.4446277
#>  [79]  4.0681398  3.8377259  3.1499073  1.3330617  1.2598803  2.7366813
#>  [85]  0.4581014  3.0132956  2.5025299  3.4539798  2.8280113  1.8014488
#>  [91]  1.4984189  4.3122491  1.1510746  3.1612537  3.3774698  4.8067410
#>  [97]  0.4342672  5.3108087  0.5127006  2.9895384
rNTS(10, 0.6, 0,1,1,0)
#>  [1] -0.16057787 -1.70232660 -1.28752082 -1.06732909 -0.74191442  0.85774365
#>  [7]  1.04173905  0.09840497 -0.36931019 -0.21420363
rNTS(10, 0.5, 1,1,1,1, NULL, "SR", 100)
#>  [1] 2.766783 2.666511 1.689549 2.066245 1.429355 3.555312 4.870261 2.578916
#>  [9] 1.474980 2.427468