Function to generate random variates of the TSS distribution.

Generates n random numbers distributed according of the tempered stable subordinator distribution.

Usage

rTSS(
  n,
  alpha = NULL,
  delta = NULL,
  lambda = NULL,
  theta = NULL,
  method = "AR",
  k = 10000
)

Arguments

n

sample size (integer).

alpha

Stability parameter. A real number between 0 and 1.

delta

Scale parameter. A real number > 0.

lambda

Tempering parameter. A real number > 0.

theta

Parameters stacked as a vector.

method

A String. Either "AR" or "SR".

k

integer: the level of truncation, if method == "SR". 10000 by default.

Value

Generates n random numbers.

Details

theta denotes the parameter vector (alpha, delta, lambda). Either provide the parameters alpha, delta, lambda individually OR provide theta. "AR" stands for the Acceptance-Rejection Method and "SR" for a truncated infinite shot noise series representation. "AR" is the standard method used. For more details, see references.

References

Massing, Till (2022), 'Parametric Estimation of Tempered Stable Laws'

Kawai, R & Masuda, H (2011), 'On simulation of tempered stable random variates' doi:10.1016/j.cam.2010.12.014

Examples

rTSS(100,0.5,1,1)
#>   [1] 0.9581645 3.2317608 2.3368535 2.0145305 2.8890137 1.5269022 1.3442685
#>   [8] 1.0875307 1.3757235 0.6293877 0.8609533 1.6038948 1.3806839 0.9405641
#>  [15] 1.8531530 0.9800251 1.7819021 2.7129422 1.8295844 1.9732573 0.7636847
#>  [22] 0.6897246 1.1490537 2.7754099 1.6337539 2.8946507 1.1781889 1.0307569
#>  [29] 1.5122251 1.6224210 0.5993133 1.7118429 0.8938606 2.7118142 1.8831639
#>  [36] 2.0452240 0.8686083 1.7632684 1.7645061 2.9923877 4.2323813 0.6374255
#>  [43] 0.7720834 3.8764465 0.7895180 1.4420274 0.9004893 1.4142230 1.0905214
#>  [50] 1.6690977 2.7651294 1.3755509 1.1275044 1.5658467 0.4965328 2.1555620
#>  [57] 2.2832768 0.9223777 3.3138413 2.3831942 0.8135419 2.3216631 4.0419958
#>  [64] 2.1884539 2.8451520 1.4101212 2.0440929 1.9875223 3.6119795 2.2810508
#>  [71] 0.7276866 0.7581314 2.2618225 1.2046871 1.0027508 1.0515658 1.6635889
#>  [78] 2.3481112 0.9097979 1.4425642 1.1443881 1.4660530 2.1618314 0.9817426
#>  [85] 0.9136776 1.6376890 3.4232746 0.7000947 2.8281407 3.3264542 2.2553393
#>  [92] 1.6679267 1.6715762 2.0881053 1.8472283 1.1583960 1.9863687 0.5316346
#>  [99] 1.2848267 0.6110433
rTSS(100,0.5,1,1,NULL,"SR",50)
#>   [1] 0.8464538 0.8898312 0.7103306 1.8796667 1.7535979 1.2215609 2.0677663
#>   [8] 0.8002452 2.1830084 1.4323540 1.0555771 1.8546975 3.5362287 2.2307076
#>  [15] 0.7830469 1.1748255 1.4895968 1.7359819 3.0114556 1.2117151 2.7284620
#>  [22] 1.6070846 1.3922310 1.4408630 2.0061423 1.4873425 1.5135238 0.5468254
#>  [29] 4.3121231 1.5490136 1.3007017 2.8627463 0.9596790 1.4416952 1.1934347
#>  [36] 0.6249925 1.7626030 2.7287875 1.2072800 1.8162459 0.8439204 2.7830714
#>  [43] 0.5888818 3.4061446 2.1245866 2.9016867 0.5318252 0.8628064 1.4974711
#>  [50] 1.2167246 2.3260400 1.0437123 1.4468855 2.2098578 0.6760700 3.2487877
#>  [57] 2.2187430 2.2004176 3.4236092 2.2170730 1.4899693 0.4726450 2.8115449
#>  [64] 0.7150540 2.1414020 0.6651097 0.5938329 2.4246610 2.9376806 0.9194805
#>  [71] 2.3564613 2.6946926 1.4234083 2.3142911 1.6375306 0.6457040 0.5986831
#>  [78] 1.2234309 2.3450759 1.0234474 1.6321601 0.8266913 0.5761497 2.0143283
#>  [85] 1.7710656 1.1623577 0.8038882 1.2269480 0.9223825 3.5486709 1.7820160
#>  [92] 4.0125738 1.3799504 1.6342694 1.9188470 2.2075202 1.9474068 1.5557131
#>  [99] 1.2351490 1.3429043