Function to generate random variates of NTS distribution.

Generates n random numbers distributed according of the normal tempered stable distribution.

rNTS(
  n,
  alpha = NULL,
  beta = NULL,
  delta = NULL,
  lambda = NULL,
  mu = NULL,
  theta = NULL,
  methodR = "AR",
  k = 10000
)

Arguments

n

sample size (integer).

alpha

A real number between 0 and 1.

beta

A gap holder.

delta

A real number > 0.

lambda

A real number > 0.

mu

A location parameter, any real number.

theta

A vector of all other arguments.

methodR

A String. Either "AR" or "SR". "AR" by default.

k

integer: the number of replications, if methodR == "SR". 10000 by default.

Value

Generates n random numbers.

Details

theta denotes the parameter vector (alpha, beta, delta, lambda, mu). Either provide the parameters individually OR provide theta. Works by a normal variance-mean mixture with a TSS distribution. Method parameter is for the method of simulating the TSS random variable, see the rTSS() function. "AR" stands for the Acceptance-Rejection Method and "SR" for a truncated infinite shot noise series representation. "AR" is the standard method used.

For more details, see references.

References

Massing, T. (2023), 'Parametric Estimation of Tempered Stable Laws'

Kawai, R & Masuda, H (2011), 'On simulation of tempered stable random variates' doi:10.1016/j.cam.2010.12.014

See also

Examples

rNTS(100, 0.5, 1,1,1,1)
#>   [1]  3.2879516  3.4101041  1.0214359  0.5515744  3.8512132  2.7507805
#>   [7]  3.6529469  1.9476845  1.4655786  3.0129511  4.0929308  2.5384294
#>  [13]  2.2887317  1.3207012  0.7209754  4.9368949  2.1627191  0.7924142
#>  [19]  2.8540381  1.1437983  1.7294241  1.1335921  1.4268728  5.0502995
#>  [25]  4.2125551  2.0832899  0.5753643  4.6611398  3.2056953  2.6268008
#>  [31]  1.9641889  3.1794679  1.4102133 -0.1230122  1.6462646  2.7558341
#>  [37]  5.2787990  3.4966902  3.7934139  0.1926948  4.0606528  1.3334007
#>  [43]  2.6123620  2.7750245  3.7776757  4.0214564  3.6841144  3.4375283
#>  [49]  2.2371163  0.6491670  2.1885719  1.8336371  7.3289288  2.4543106
#>  [55]  0.7099233  3.5205371  1.2692195  2.3763285  3.0701399  1.3757288
#>  [61]  3.8756806  2.5630446  1.1737934  3.1481858  4.3406914  3.6169432
#>  [67]  4.0385188  1.2894373  1.2521646  5.1064433  1.9226711  0.7746370
#>  [73]  1.0722934  2.4278012  6.3488198  5.9645826  2.3303689  4.3508408
#>  [79]  1.5670902  3.0531305  2.3534948  4.7553563  3.6299442  1.0271777
#>  [85]  1.3439921  3.8808479  3.5716155  3.9844620  3.0650451  1.9070459
#>  [91]  4.8898528  2.3921113  2.5995507 -0.5105068 -0.3709426  2.9444020
#>  [97]  1.5605682  0.9041690  1.7927110  2.1013320
rNTS(10, 0.6, 0,1,1,0)
#>  [1] -2.30137010  1.65541957  0.73357680  2.29349955 -0.45165344  0.71563534
#>  [7]  0.20096949  1.30059174  0.06681498  2.50643812
rNTS(10, 0.5, 1,1,1,1, NULL, "SR", 100)
#>  [1] 1.624711 2.650995 1.478307 2.573775 2.787875 2.434197 3.137854 0.553372
#>  [9] 3.792414 2.907259