Function to generate random variates of NTS distribution.

Generates n random numbers distributed according of the normal tempered stable distribution.

rNTS(
  n,
  alpha = NULL,
  beta = NULL,
  delta = NULL,
  lambda = NULL,
  mu = NULL,
  theta = NULL,
  methodR = "AR",
  k = 10000
)

Arguments

n

sample size (integer).

alpha

A real number between 0 and 1.

beta

A gap holder.

delta

A real number > 0.

lambda

A real number > 0.

mu

A location parameter, any real number.

theta

A vector of all other arguments.

methodR

A String. Either "AR" or "SR". "AR" by default.

k

integer: the number of replications, if methodR == "SR". 10000 by default.

Value

Generates n random numbers.

Details

theta denotes the parameter vector (alpha, beta, delta, lambda, mu). Either provide the parameters individually OR provide theta. Works by a normal variance-mean mixture with a TSS distribution. Method parameter is for the method of simulating the TSS random variable, see the rTSS() function. "AR" stands for the Acceptance-Rejection Method and "SR" for a truncated infinite shot noise series representation. "AR" is the standard method used.

For more details, see references.

References

Massing, T. (2023), 'Parametric Estimation of Tempered Stable Laws'

Kawai, R & Masuda, H (2011), 'On simulation of tempered stable random variates' doi:10.1016/j.cam.2010.12.014

See also

Examples

rNTS(100, 0.5, 1,1,1,1)
#>   [1]  1.5712849  3.3636296  2.4740426  1.5215607  2.0658730  0.4807176
#>   [7]  0.7406872  2.4997099  2.5857043  2.8491652 -0.2358095  3.1666370
#>  [13]  1.0752192  2.1868949  2.8364707  1.7243228  1.1922663  1.5751490
#>  [19] -0.3167525  1.9710819 -0.1957847  4.7997657  3.3248553  1.4995481
#>  [25] -0.2447276  1.6288385  0.3170280 -0.1327327  4.3402985  2.7004308
#>  [31]  3.4632336  3.0269011  6.0459818  1.2117190  2.7880124  4.2405516
#>  [37]  3.5703224  2.4220313  2.8450260 -0.2452927  0.1356603  2.4938152
#>  [43]  0.5389849  0.8782596  2.8527085  3.9477392  1.1046180  4.5820824
#>  [49]  3.2315282  0.9278475  2.7158734  1.9341037  6.3430118  2.1328378
#>  [55]  3.8006981  3.6448575  0.2494114  2.0932060  3.6112306  0.3421005
#>  [61]  1.8984201  6.4682684  2.8564083  3.0981928  1.8527094  2.5233967
#>  [67]  3.5156693  1.0986529  2.1363264  2.9883262  3.1504111  2.4303381
#>  [73]  4.5263933  2.8400382  4.9971386  2.0981375  1.6673353  7.3503265
#>  [79]  1.9798463  4.1344922  1.8737605  2.5236578  3.4956309  4.5539077
#>  [85] -0.3102539  2.9336845  1.8932272  0.5349852  1.9146472  2.9524863
#>  [91]  2.3078019  0.8541859  3.3793188  0.9168764  0.8307787  5.2549604
#>  [97] -0.3446385  0.8357621  4.5551545  2.8064830
rNTS(10, 0.6, 0,1,1,0)
#>  [1] -1.2029884 -2.3903645  0.4151353 -0.6232252 -1.4448662 -2.0968086
#>  [7]  0.1934787  0.4232655 -2.2662249  0.3898931
rNTS(10, 0.5, 1,1,1,1, NULL, "SR", 100)
#>  [1] 3.64822144 0.02420164 1.19749873 1.49062579 3.19217803 1.34502080
#>  [7] 4.24344013 1.72546722 1.61571815 2.48972600