Function to generate random variates of NTS distribution.

Generates n random numbers distributed according of the normal tempered stable distribution.

rNTS(
  n,
  alpha = NULL,
  beta = NULL,
  delta = NULL,
  lambda = NULL,
  mu = NULL,
  theta = NULL,
  methodR = "TM",
  k = 10000
)

Arguments

n

sample size (integer).

alpha

A real number between 0 and 1.

beta

A gap holder.

delta

A real number > 0.

lambda

A real number > 0.

mu

A location parameter, any real number.

theta

A vector of all other arguments.

methodR

A String. Either "TM","AR" or "SR". "TM" by default.

k

integer: the number of replications, if methodR == "SR". 10000 by default.

Value

Generates n random numbers.

Details

theta denotes the parameter vector (alpha, beta, delta, lambda, mu). Either provide the parameters individually OR provide theta. Works by a normal variance-mean mixture with a TSS distribution. Method parameter is for the method of simulating the TSS random variable, see the rTSS() function. "AR" stands for the Acceptance-Rejection Method and "SR" for a truncated infinite shot noise series representation. "TM" stands for Two Methods as two different methods are used depending on which will be faster. In this method the function copula::retstable() is called. "TM" is the standard method used. For more details, see references.

It is recommended to check the generated random numbers once for each distribution using the density function. If the random numbers are shifted, e.g. for the method "SR", it may be worthwhile to increase k.

For more details, see references.

References

Massing, T. (2023), 'Parametric Estimation of Tempered Stable Laws'

Kawai, R & Masuda, H (2011), 'On simulation of tempered stable random variates' doi:10.1016/j.cam.2010.12.014

Hofert, M (2011), 'Sampling Exponentially Tilted Stable Distributions' doi:10.1145/2043635.2043638

See also

Examples

rNTS(100, 0.5, 1,1,1,1)
#>   [1] 1.4636083 2.6360423 2.1766965 2.8018843 1.7342307 3.5982051 1.1997908
#>   [8] 0.8817551 3.5852330 3.4733878 1.5512980 4.4679725 3.6268453 4.8419036
#>  [15] 2.1593173 1.6919549 5.7713791 2.8254973 2.6863610 2.2535301 2.4680277
#>  [22] 4.9493125 2.1186116 2.0099785 2.9081829 1.9905712 3.9407426 3.4965645
#>  [29] 1.0363117 3.0509498 2.8947659 0.4155517 1.4235055 3.5999523 2.5080810
#>  [36] 2.5773298 6.4051235 3.7315096 0.2876812 4.3959787 2.2859749 4.1320620
#>  [43] 2.9200547 5.6731849 2.6967121 1.9485424 5.7228173 2.8614948 5.4893452
#>  [50] 4.3045866 1.6473618 1.2357617 1.1869528 3.0040115 2.5842960 1.6113567
#>  [57] 3.1402387 1.1752424 0.7647568 7.0720340 2.4941040 4.0239919 2.9404714
#>  [64] 3.2472033 6.2588108 1.0407243 4.1482970 0.4415743 1.4601052 2.1700673
#>  [71] 2.4121521 2.4452373 3.4418356 3.1090939 0.5519724 6.7709336 4.2731130
#>  [78] 3.1117790 3.2895577 2.9405255 4.3803703 3.3655892 4.1812501 2.6459271
#>  [85] 2.1954582 1.8526715 3.5460767 4.9119569 3.0533405 4.5999560 0.1958741
#>  [92] 2.8429981 1.9262387 0.3886232 1.1819131 3.9491086 2.4084257 2.2543371
#>  [99] 2.9718048 2.9166128
rNTS(10, 0.6, 0,1,1,0)
#>  [1] -1.2116797 -0.2504829  0.9804680 -0.1596497 -1.8791735 -1.5570253
#>  [7]  0.9426611 -2.9588112 -3.3739887  1.9994552
rNTS(10, 0.5, 1,1,1,1, NULL, "SR", 100)
#>  [1] 0.4332245 3.3458988 1.9613651 2.1938192 3.1474604 4.5964612 4.1002525
#>  [8] 4.8025085 2.0123186 4.3384895