Function to generate random variates of NTS distribution.

Generates n random numbers distributed according of the normal tempered stable distribution.

rNTS(
  n,
  alpha = NULL,
  beta = NULL,
  delta = NULL,
  lambda = NULL,
  mu = NULL,
  theta = NULL,
  methodR = "AR",
  k = 10000
)

Arguments

n

sample size (integer).

alpha

A real number between 0 and 1.

beta

A gap holder.

delta

A real number > 0.

lambda

A real number > 0.

mu

A location parameter, any real number.

theta

A vector of all other arguments.

methodR

A String. Either "AR" or "SR". "AR" by default.

k

integer: the number of replications, if methodR == "SR". 10000 by default.

Value

Generates n random numbers.

Details

theta denotes the parameter vector (alpha, beta, delta, lambda, mu). Either provide the parameters individually OR provide theta. Works by a normal variance-mean mixture with a TSS distribution. Method parameter is for the method of simulating the TSS random variable, see the rTSS() function. "AR" stands for the Acceptance-Rejection Method and "SR" for a truncated infinite shot noise series representation. "AR" is the standard method used.

For more details, see references.

References

Massing, T. (2023), 'Parametric Estimation of Tempered Stable Laws'

Kawai, R & Masuda, H (2011), 'On simulation of tempered stable random variates' doi:10.1016/j.cam.2010.12.014

See also

Examples

rNTS(100, 0.5, 1,1,1,1)
#>   [1]  2.45078734 -0.01187656  2.97304011  1.15707729  2.68330861  2.06677260
#>   [7]  0.64243339  1.55510578  2.41311540  2.67979260  1.59617457  3.42103249
#>  [13]  7.34443773  5.50038198  3.00199404  4.11666364  0.92832620  3.03755753
#>  [19]  1.34953864  1.34961050  3.77609702  2.15638656  3.80142827  0.51447018
#>  [25]  1.93637780  4.67833502  3.23926911  1.46146869  2.54543471  2.75684064
#>  [31]  2.42863310  1.97824203  1.59869998  3.37917179  5.60297670  1.47969653
#>  [37]  5.86318673  2.46458643  3.93014135  4.09226997  3.78527849  3.51067072
#>  [43]  4.43222373  1.24612846  3.47236918  3.68967292  1.24861801  3.74419367
#>  [49] -0.42852075  5.25991588  2.02253165  3.01261040  1.61605755  2.89869595
#>  [55]  3.26098579  1.42309422  4.26880851  3.70644992  4.68144522  1.94167009
#>  [61]  2.38574642  4.10815452  4.45271889  4.13593774  1.45140396  1.31508141
#>  [67]  2.35604005  1.54041409  0.76579597  1.72516234 -0.53849811  3.03143558
#>  [73]  1.40981180  3.76419559  0.75957505  2.26246180  2.44160653  3.75770294
#>  [79]  2.17204791  2.55369556  2.77304791  3.33365911  2.16835233  2.11453855
#>  [85]  3.37462265  5.29408690  4.15840402  2.24879236  3.22635662  1.87788155
#>  [91]  1.87907437  1.28889980  2.41072648  1.76331181  2.13127305  4.20937749
#>  [97]  1.80494313  0.78985045  1.36334688  2.37629810
rNTS(10, 0.6, 0,1,1,0)
#>  [1] -0.2958980 -0.4552026 -0.1509250 -2.1595336 -2.1581418 -2.7581151
#>  [7]  0.1277273  3.9152752 -0.7632020  0.9171643
rNTS(10, 0.5, 1,1,1,1, NULL, "SR", 100)
#>  [1] 1.8921911 1.4437624 3.9247784 3.7862710 1.8282949 0.8142112 6.9351370
#>  [8] 5.1239174 2.3816976 0.9328438