Function to generate random variates of NTS distribution.

Generates n random numbers distributed according of the normal tempered stable distribution.

rNTS(
  n,
  alpha = NULL,
  beta = NULL,
  delta = NULL,
  lambda = NULL,
  mu = NULL,
  theta = NULL,
  methodR = "AR",
  k = 10000
)

Arguments

n

sample size (integer).

alpha

A real number between 0 and 1.

beta

A gap holder.

delta

A real number > 0.

lambda

A real number > 0.

mu

A location parameter, any real number.

theta

A vector of all other arguments.

methodR

A String. Either "AR" or "SR". "AR" by default.

k

integer: the number of replications, if methodR == "SR". 10000 by default.

Value

Generates n random numbers.

Details

theta denotes the parameter vector (alpha, beta, delta, lambda, mu). Either provide the parameters individually OR provide theta. Works by a normal variance-mean mixture with a TSS distribution. Method parameter is for the method of simulating the TSS random variable, see the rTSS() function. "AR" stands for the Acceptance-Rejection Method and "SR" for a truncated infinite shot noise series representation. "AR" is the standard method used.

For more details, see references.

References

Massing, Till (2022), 'Parametric Estimation of Tempered Stable Laws'

Kawai, R & Masuda, H (2011), 'On simulation of tempered stable random variates' doi:10.1016/j.cam.2010.12.014

See also

Examples

rNTS(100, 0.5, 1,1,1,1)
#>   [1]  5.3434493  4.9858705  1.3478498  3.1475256  5.9818665  3.4317441
#>   [7]  2.1861134  4.9618035  1.6860847  2.4139745  4.2795827  3.0262441
#>  [13]  2.0353548  1.2695151 12.9428972  2.5884570  6.9230454  3.2782379
#>  [19]  3.9552521  4.2328986  1.3159494  2.6749865  1.7277617  1.1917504
#>  [25]  3.6089731  0.9678045  0.9299114  2.4874676  2.4942463  2.1433577
#>  [31]  1.5637793  0.7227632  2.4282084  3.7107611  1.2184819  2.6160476
#>  [37] -0.1537764  3.3134987  3.7115563  0.8307318  2.6334710  7.8550407
#>  [43]  2.4055955  3.7452439 -0.4762264  3.0440655  3.3561056  2.4322052
#>  [49]  4.6578336  0.5967566  7.4533441  3.3697800  2.2553502  3.0813549
#>  [55]  3.5137019  0.9768484 -0.6091346  3.0898385  0.1154188  2.5479603
#>  [61]  1.3585985  2.1996881  3.8898904  2.6851847  3.9228388  1.2234445
#>  [67]  3.3699485  2.7689962  2.4666731  0.0956665  4.9871685  3.2544729
#>  [73]  2.5235968  4.4082594  1.8897920  1.0068866  1.4028615  3.8009287
#>  [79]  1.2682385  1.3983001  2.0688723  4.1443258  4.4821084  3.8433799
#>  [85]  1.8165082  4.1151836  5.5566507  2.4041641  3.7080387  0.9256835
#>  [91]  3.2557290  1.9538979  1.3112987  2.9305196  1.5110511  1.5542273
#>  [97]  0.8850376  4.8672389  5.6158949  6.4913356
rNTS(10, 0.6, 0,1,1,0)
#>  [1] -0.31785915 -1.97417033 -1.37958456  1.50220959 -1.19047607 -2.12716150
#>  [7] -1.25104568 -1.61695138  0.03645815 -2.88772235
rNTS(10, 0.5, 1,1,1,1, NULL, "SR", 100)
#>  [1] 6.651663 3.988590 4.911222 6.429533 4.427151 1.277080 2.788756 4.217246
#>  [9] 0.879107 3.368303