Function to generate random variates of NTS distribution.

Generates n random numbers distributed according of the normal tempered stable distribution.

rNTS(
  n,
  alpha = NULL,
  beta = NULL,
  delta = NULL,
  lambda = NULL,
  mu = NULL,
  theta = NULL,
  methodR = "AR",
  k = 10000
)

Arguments

n

sample size (integer).

alpha

A real number between 0 and 1.

beta

A gap holder.

delta

A real number > 0.

lambda

A real number > 0.

mu

A location parameter, any real number.

theta

A vector of all other arguments.

methodR

A String. Either "AR" or "SR". "AR" by default.

k

integer: the number of replications, if methodR == "SR". 10000 by default.

Value

Generates n random numbers.

Details

theta denotes the parameter vector (alpha, beta, delta, lambda, mu). Either provide the parameters individually OR provide theta. Works by a normal variance-mean mixture with a TSS distribution. Method parameter is for the method of simulating the TSS random variable, see the rTSS() function. "AR" stands for the Acceptance-Rejection Method and "SR" for a truncated infinite shot noise series representation. "AR" is the standard method used.

For more details, see references.

References

Massing, Till (2022), 'Parametric Estimation of Tempered Stable Laws'

Kawai, R & Masuda, H (2011), 'On simulation of tempered stable random variates' doi:10.1016/j.cam.2010.12.014

See also

Examples

rNTS(100, 0.5, 1,1,1,1)
#>   [1]  2.4097849  2.0109142  1.9427488  1.3333730  0.9293710  4.5951178
#>   [7]  1.4245488  2.6488370  1.7130700  1.8318579  3.2668969  2.5379290
#>  [13]  3.4936755  2.1192784  1.2588445  5.3157956  5.9797371  3.2351227
#>  [19]  2.2364534  5.4809484  2.3482045  1.2852553  2.8410331  1.9861807
#>  [25]  1.2799825  6.0819281  1.0054089  2.0839713  2.1618235  1.6992945
#>  [31]  3.6121774  3.4582746  2.6892327  1.5110721  3.3547086  1.3929884
#>  [37]  4.7848328  1.6498846  1.7255599  3.6807035  1.9564286  2.9681580
#>  [43] -0.4880775  5.8231705  2.9921096  4.3461432  4.0407210  3.1386792
#>  [49]  0.6225175  1.7359392  3.9146015  4.2966641  5.2712302  2.9307218
#>  [55]  1.8578398  5.3547142  1.6405122  3.9896363  2.7129289  3.1343783
#>  [61]  2.4630167  2.7710671  1.6744676  8.8618380  0.7314872  3.3279285
#>  [67]  2.7012028  2.0786565  0.7863519 -0.1617999  2.5170793  0.3208225
#>  [73]  6.7881714  3.8161264  1.4152615  2.4536522  2.9215743  5.3196702
#>  [79]  3.5429607  1.6179957  3.5439688  2.6060967  9.6625382  3.4201105
#>  [85]  2.6331269  3.0466504  6.6779100  5.9080420  1.8451202  1.1034179
#>  [91]  1.5271763  3.8543659  3.0969405  4.6029457  2.3224390  3.4530776
#>  [97]  2.7234148  2.6567097  2.4241349  1.6758761
rNTS(10, 0.6, 0,1,1,0)
#>  [1]  0.11888504  0.84441807  1.37711343  4.44397481  0.05624685 -0.12178773
#>  [7] -1.29653350 -1.45603616 -0.91902022  1.33650436
rNTS(10, 0.5, 1,1,1,1, NULL, "SR", 100)
#>  [1] 1.4949399 1.1940259 1.5248849 1.5883088 0.9297839 1.4562239 2.6797152
#>  [8] 3.3685523 2.8899436 4.4650205