Function to generate random variates of NTS distribution.

Generates n random numbers distributed according of the normal tempered stable distribution.

rNTS(
  n,
  alpha = NULL,
  beta = NULL,
  delta = NULL,
  lambda = NULL,
  mu = NULL,
  theta = NULL,
  method = "AR",
  k = 10000
)

Arguments

n

sample size (integer).

alpha

A real number between 0 and 1.

beta

A gap holder.

delta

A real number > 0.

lambda

A real number > 0.

mu

A location parameter, any real number.

theta

A vector of all other arguments.

method

A String. Either "AR" or "SR". "AR" by default.

k

integer: the number of replications, if method == "SR". 10000 by default.

Value

Generates n random numbers.

Details

theta denotes the parameter vector (alpha, beta, delta, lambda, mu). Either provide the parameters individually OR provide theta. Works by a normal variance-mean mixture with a TSS distribution. Method parameter is for the method of simulating the TSS random variable, see the rTSS() function. "AR" stands for the Acceptance-Rejection Method and "SR" for a truncated infinite shot noise series representation. "AR" is the standard method used.

For more details, see references.

References

Massing, Till (2022), 'Parametric Estimation of Tempered Stable Laws'

Kawai, R & Masuda, H (2011), 'On simulation of tempered stable random variates' doi:10.1016/j.cam.2010.12.014

See also

Examples

rNTS(100, 0.5, 1,1,1,1)
#>   [1]  0.05533230  0.46388018  2.65274245  4.49459084  2.08216862  2.31643140
#>   [7]  2.47703458  3.05021499  2.52564447  1.55895609  4.70725895  1.35597296
#>  [13]  4.34655965  1.95892797  0.90635217  2.21924643 -0.20306493  2.54292101
#>  [19]  2.17446262  2.78700346  2.64408906  0.77083588  2.75849154  0.53486566
#>  [25]  5.09023634  5.38823211  1.70299408  1.77571631  1.97840938  2.28815802
#>  [31]  2.89337733  0.40562870  3.92503246  1.93497175  1.85197773  3.06731985
#>  [37]  4.47501676  2.68642922  0.22293809  1.22212357  3.07106490  3.93456751
#>  [43]  1.35207909  3.47972273  1.49372769  4.23661733  5.01698153  3.29799344
#>  [49]  4.01046932  0.94260592  1.59431850  0.87355201  3.90931978  3.14362602
#>  [55]  2.10175718  2.59896304  2.00851416  1.78730964  6.30458841  3.73486306
#>  [61]  1.90490901  3.70181974  1.58490815  0.77491090  3.67076250  0.07709958
#>  [67]  1.73487326  3.55756639  2.02779411  3.76028149  0.19964503  1.22725809
#>  [73]  4.52989348  2.45901016  2.86349526  4.83306386  6.22420600 -0.25141996
#>  [79]  1.10281795  1.79768793  5.17840965  1.44597021  2.74200410  5.86538588
#>  [85]  0.64901428  1.96659976  2.73718428  0.48790478  4.37645884  5.49010447
#>  [91]  2.12263426  5.69315102  2.89542344  1.23385379  5.45740275  3.14313498
#>  [97]  3.45980277  2.30850229  3.93129485  2.25340529
rNTS(10, 0.6, 0,1,1,0)
#>  [1] -1.0839185 -0.7760522  0.4953347 -0.1350433  1.3358755  2.2646726
#>  [7]  1.3281321  1.2739373  0.7885048 -0.2609306
rNTS(10, 0.5, 1,1,1,1, NULL, "SR", 100)
#>  [1] 1.707155 2.695923 1.459443 4.379673 3.433056 1.592940 1.316619 4.173866
#>  [9] 2.296499 1.438714