Function to generate random variates of NTS distribution.

Generates n random numbers distributed according of the normal tempered stable distribution.

Usage

rNTS(
  n,
  alpha = NULL,
  beta = NULL,
  delta = NULL,
  lambda = NULL,
  mu = NULL,
  theta = NULL,
  method = "AR",
  k = 10000
)

Arguments

n

sample size (integer).

alpha

A real number between 0 and 1.

beta

A gap holder.

delta

A real number > 0.

lambda

A real number > 0.

mu

A location parameter, any real number.

theta

A vector of all other arguments.

method

A String. Either "AR" or "SR". "AR" by default.

k

integer: the number of replications, if method == "SR". 10000 by default.

Value

Generates n random numbers.

Details

theta denotes the parameter vector (alpha, beta, delta, lambda, mu). Either provide the parameters individually OR provide theta. Works by a normal variance-mean mixture with a TSS distribution. Method parameter is for the method of simulating the TSS random variable, see the rTSS() function. "AR" stands for the Acceptance-Rejection Method and "SR" for a truncated infinite shot noise series representation. "AR" is the standard method used.

For more details, see references.

References

Massing, Till (2022), 'Parametric Estimation of Tempered Stable Laws'

Kawai, R & Masuda, H (2011), 'On simulation of tempered stable random variates' doi:10.1016/j.cam.2010.12.014

See also

Examples

rNTS(100, 0.5, 1,1,1,1)
#>   [1] 4.3526493 1.3794969 3.0497784 1.2148422 2.5667551 2.4350292 1.7662197
#>   [8] 0.4647955 1.9416852 2.6329902 2.8324597 4.7115791 1.0431679 3.1731380
#>  [15] 2.0768677 3.0246277 3.6349789 6.7133249 4.6370542 1.2369523 1.2618045
#>  [22] 2.1540106 3.4152425 0.4522202 3.9765363 1.9265241 1.2567702 3.8123123
#>  [29] 4.4342025 1.2034252 1.7990951 3.3848335 2.9887389 3.9875690 2.1847246
#>  [36] 3.1716842 2.6999303 5.1464669 0.8218184 1.1549964 2.3101889 1.5516743
#>  [43] 0.6178350 2.1528287 0.6738918 2.5560529 4.2160864 3.0471116 5.7733261
#>  [50] 1.3119856 3.3247323 0.3175360 6.6012846 5.2543427 4.8424450 2.3834910
#>  [57] 2.5941751 0.8147093 2.8799476 2.2735667 2.5276994 2.0340919 4.1206813
#>  [64] 2.5054824 5.0002856 3.4542397 4.7437843 3.7663133 2.0029094 1.6432171
#>  [71] 0.9667719 1.5475633 3.3008093 1.8475159 3.7610036 1.6720224 3.7610541
#>  [78] 5.0297725 2.2904559 0.9016524 2.7165352 3.7409524 2.1095574 1.3012788
#>  [85] 1.5707984 0.6276293 2.4493777 3.5329658 3.0367325 3.1315447 1.1333410
#>  [92] 4.3737544 1.7090450 3.2192907 3.7178144 1.8536418 2.8370245 5.8588189
#>  [99] 4.6083851 3.2849985
rNTS(10, 0.6, 0,1,1,0)
#>  [1]  1.04858948 -1.04821843 -1.05739742 -1.40484790 -0.33840134  0.16368084
#>  [7]  1.42824297  1.13395765 -0.07034087  1.18914525
rNTS(10, 0.5, 1,1,1,1, NULL, "SR", 100)
#>  [1] 2.9983607 1.2261244 2.3231206 0.4032546 4.3540289 6.6679312 3.5932652
#>  [8] 1.3626113 3.6560635 5.3593587