Function to generate random variates of NTS distribution.

Generates n random numbers distributed according of the normal tempered stable distribution.

rNTS(
  n,
  alpha = NULL,
  beta = NULL,
  delta = NULL,
  lambda = NULL,
  mu = NULL,
  theta = NULL,
  methodR = "AR",
  k = 10000
)

Arguments

n

sample size (integer).

alpha

A real number between 0 and 1.

beta

A gap holder.

delta

A real number > 0.

lambda

A real number > 0.

mu

A location parameter, any real number.

theta

A vector of all other arguments.

methodR

A String. Either "AR" or "SR". "AR" by default.

k

integer: the number of replications, if methodR == "SR". 10000 by default.

Value

Generates n random numbers.

Details

theta denotes the parameter vector (alpha, beta, delta, lambda, mu). Either provide the parameters individually OR provide theta. Works by a normal variance-mean mixture with a TSS distribution. Method parameter is for the method of simulating the TSS random variable, see the rTSS() function. "AR" stands for the Acceptance-Rejection Method and "SR" for a truncated infinite shot noise series representation. "AR" is the standard method used.

For more details, see references.

References

Massing, Till (2022), 'Parametric Estimation of Tempered Stable Laws'

Kawai, R & Masuda, H (2011), 'On simulation of tempered stable random variates' doi:10.1016/j.cam.2010.12.014

See also

Examples

rNTS(100, 0.5, 1,1,1,1)
#>   [1]  3.2688515  2.2699757  1.5088691  5.0070987  2.7857515  0.8681505
#>   [7]  1.8551829  2.2891190  1.0385637  1.2255602  1.7996334  4.3318395
#>  [13]  9.4144293  2.8992063  1.8142270  1.4409684  1.6936796  3.4867304
#>  [19]  2.5460144  2.6749266  3.0118320  0.3944926  1.4434352  1.3735846
#>  [25]  6.3490662  8.9887964  2.1622310  2.2355335  0.5418792  2.8975686
#>  [31]  4.9368972  2.6969820  2.3588217  2.6636489  2.9911710  2.4486420
#>  [37] -0.1355034  1.2136839  1.5314995  3.0407358  2.7343659  2.7696611
#>  [43]  4.3963239  4.0649081  0.8695117  3.9385475  2.7838476  3.2288998
#>  [49]  1.7420743  1.5332313  2.1975184  4.5667226  5.2971568  2.1523488
#>  [55]  2.6299880  0.5502991  1.8227754  0.6147767  1.8430345  1.3026074
#>  [61]  6.8211731  3.2050355  0.7961423  2.7958885  2.6859863  2.7783101
#>  [67]  0.8805525  1.9806230  2.2079540  3.9689505  2.0244718  1.1034207
#>  [73]  4.4517538  3.4818997  2.1472047  1.3137639  1.0083367  3.0095255
#>  [79]  5.3743679  1.9511713  2.3975721  4.1311787  1.9157777  1.3891012
#>  [85]  1.2897530  2.8904589  1.8245062  1.5259215  3.0564295  2.0592870
#>  [91]  2.3893930  3.9605501  5.9062674  1.9399458  3.9539661  1.2154221
#>  [97]  2.7314711 -0.2659995  1.4488496  2.8980956
rNTS(10, 0.6, 0,1,1,0)
#>  [1] -0.629699417 -0.003174534 -0.770408298  1.724174344  0.205921806
#>  [6]  1.062798473  0.621381335 -1.966660241 -2.336620358 -1.423056230
rNTS(10, 0.5, 1,1,1,1, NULL, "SR", 100)
#>  [1] 2.210593 2.230220 3.147782 3.052377 4.055470 3.375035 5.837272 3.698882
#>  [9] 3.901756 2.374324