Function to generate random variates of NTS distribution.

Generates n random numbers distributed according of the normal tempered stable distribution.

rNTS(
  n,
  alpha = NULL,
  beta = NULL,
  delta = NULL,
  lambda = NULL,
  mu = NULL,
  theta = NULL,
  methodR = "AR",
  k = 10000
)

Arguments

n

sample size (integer).

alpha

A real number between 0 and 1.

beta

A gap holder.

delta

A real number > 0.

lambda

A real number > 0.

mu

A location parameter, any real number.

theta

A vector of all other arguments.

methodR

A String. Either "AR" or "SR". "AR" by default.

k

integer: the number of replications, if methodR == "SR". 10000 by default.

Value

Generates n random numbers.

Details

theta denotes the parameter vector (alpha, beta, delta, lambda, mu). Either provide the parameters individually OR provide theta. Works by a normal variance-mean mixture with a TSS distribution. Method parameter is for the method of simulating the TSS random variable, see the rTSS() function. "AR" stands for the Acceptance-Rejection Method and "SR" for a truncated infinite shot noise series representation. "AR" is the standard method used.

For more details, see references.

References

Massing, Till (2022), 'Parametric Estimation of Tempered Stable Laws'

Kawai, R & Masuda, H (2011), 'On simulation of tempered stable random variates' doi:10.1016/j.cam.2010.12.014

See also

Examples

rNTS(100, 0.5, 1,1,1,1)
#>   [1]  3.4904193  5.9431509  4.5634337  2.1174863  2.7706177  3.6709787
#>   [7]  2.9705611  3.4522113  1.0468999  2.9675688  2.5943476  3.2811287
#>  [13]  2.1480500  3.5247386  1.8529067  3.8226529  2.7587417  1.7230027
#>  [19]  2.9173858  2.8593691  2.5115106  2.3262728  1.8288129  1.5227047
#>  [25]  1.9502849  3.5494692  3.0340852  1.1639826  3.9017360  3.0369028
#>  [31]  0.9656511  3.8610688  2.9242135  2.8330924  3.0443856 -0.2161404
#>  [37]  2.3626760  3.0244616  0.8463324  0.7200955  4.7440536  2.5682344
#>  [43]  3.2702152  3.2961326  4.6286816  1.1329659  1.8665613  0.1717705
#>  [49]  8.4802081  3.0911481  2.4257426  2.8341576  2.6680903  1.9495628
#>  [55]  0.9998896  2.7869870  3.0860426  5.7100848  1.7770833  3.3531307
#>  [61]  3.4096909  2.9610401  1.2366132  1.2683042  0.7640391  2.6457599
#>  [67]  2.2621338  2.6692661  3.2009347  1.3319624  3.7393893  2.7051509
#>  [73]  3.2475281  2.1258351  2.1789233  0.9058303  3.6716893  0.8930669
#>  [79]  4.1666279  0.8341212  1.0865771  2.7998270  2.7605834  2.4907893
#>  [85]  0.8057941  0.1145367  2.5639686  2.1764439  2.0159046  3.1080604
#>  [91]  1.9061890  2.0472686  4.5286788  3.3238653  3.7563700  2.0608545
#>  [97]  1.2550249  2.7699320  1.7011671  0.9978634
rNTS(10, 0.6, 0,1,1,0)
#>  [1]  1.1594469  0.8842630  1.8691250 -0.1249010  1.8395036  1.7417994
#>  [7]  0.7666264  1.0940996 -1.8978522  1.2594814
rNTS(10, 0.5, 1,1,1,1, NULL, "SR", 100)
#>  [1] 1.831907 3.866511 2.397166 2.271924 1.935243 5.438435 4.756782 4.995528
#>  [9] 4.658126 1.185406