Function to generate random variates of the TSS distribution.

Generates n random numbers distributed according of the tempered stable subordinator distribution.

rTSS(
  n,
  alpha = NULL,
  delta = NULL,
  lambda = NULL,
  theta = NULL,
  methodR = "AR",
  k = 10000
)

Arguments

n

sample size (integer).

alpha

Stability parameter. A real number between 0 and 1.

delta

Scale parameter. A real number > 0.

lambda

Tempering parameter. A real number > 0.

theta

Parameters stacked as a vector.

methodR

A String. Either "AR" or "SR".

k

integer: the level of truncation, if methodR == "SR". 10000 by default.

Value

Generates n random numbers.

Details

theta denotes the parameter vector (alpha, delta, lambda). Either provide the parameters alpha, delta, lambda individually OR provide theta. "AR" stands for the Acceptance-Rejection Method and "SR" for a truncated infinite shot noise series representation. "AR" is the standard method used. For more details, see references.

References

Massing, T. (2023), 'Parametric Estimation of Tempered Stable Laws'

Kawai, R & Masuda, H (2011), 'On simulation of tempered stable random variates' doi:10.1016/j.cam.2010.12.014

Examples

rTSS(100,0.5,1,1)
#>   [1] 1.9049988 3.9191187 2.0503790 1.8911817 1.2884296 1.1965656 2.0363673
#>   [8] 5.1150451 0.9563463 0.7644210 2.3547096 1.2814746 4.1696949 2.3136563
#>  [15] 2.2649104 1.6972174 1.1259181 1.5701278 1.5740003 1.3983354 1.0225786
#>  [22] 1.1117760 2.4072141 1.2390590 2.4351540 1.5461451 1.4502372 2.7539457
#>  [29] 2.0711859 1.4026413 1.0567657 3.4874146 1.2254821 2.8137614 2.4697923
#>  [36] 1.9585292 2.2615981 2.6006709 1.8975043 2.2044178 1.7850773 0.8675857
#>  [43] 2.1481535 1.5546694 1.8000522 2.9555562 0.9762566 2.8580272 1.6680197
#>  [50] 0.7994765 2.3786442 2.2623827 1.4499651 2.0993855 3.5819988 0.6807093
#>  [57] 0.8971232 1.6104360 1.1237487 1.0274414 1.9794621 0.6079061 1.5738203
#>  [64] 0.8233529 2.1743005 1.0497219 1.6117037 2.1333783 1.5276908 1.7954776
#>  [71] 1.9287823 2.1122639 2.3643555 1.0540583 0.8250562 2.4199151 1.4293052
#>  [78] 0.7431755 1.9479194 0.8221491 0.8389079 0.6657993 4.9931133 3.5351976
#>  [85] 0.6213913 1.8099990 2.5592076 0.6936954 2.2739916 1.4780423 0.9511469
#>  [92] 1.6466845 2.1239573 1.1268928 1.2189754 1.6686868 1.1557531 1.8440102
#>  [99] 2.1356447 1.8877571
rTSS(100,0.5,1,1,NULL,"SR",50)
#>   [1] 0.8689151 0.9220306 1.4394159 1.3101097 2.0102508 1.8551380 1.9410681
#>   [8] 1.0676643 1.5133521 4.0175155 0.9626440 0.8116574 4.3083956 0.9357273
#>  [15] 1.0102650 0.3748553 2.3250406 2.7628686 1.5310185 1.0826456 3.5729415
#>  [22] 1.5626143 3.4860557 3.4352367 3.3658600 1.2508768 2.0857676 0.8244685
#>  [29] 2.2452902 0.8975596 2.1493600 1.4119845 1.0380866 2.2066825 1.3537179
#>  [36] 0.5637728 1.9937824 2.9014240 0.2568950 1.8809076 1.0053037 2.0019854
#>  [43] 1.2011497 2.2485284 0.6063934 1.1058497 1.1993255 0.7818216 0.6685430
#>  [50] 1.9835122 0.3147637 3.1692768 0.4949140 2.1862952 0.9618522 1.8304669
#>  [57] 1.3962471 0.9763217 1.1947846 1.9697476 3.0372623 2.1260579 3.6392898
#>  [64] 0.8849197 1.3894549 1.8822885 1.6463445 0.9161112 0.7950216 2.3101672
#>  [71] 2.8059921 1.8014170 1.6970276 1.6476899 2.0069430 0.9022687 2.5822544
#>  [78] 3.3116228 1.8114276 1.5068913 0.6165315 2.2271372 1.6232266 2.2136696
#>  [85] 1.9069805 1.5036046 1.8441288 1.4710790 1.6998778 2.2677377 0.7994297
#>  [92] 0.4394420 1.7926713 1.3654756 1.1231483 2.5756848 0.4332075 1.5739913
#>  [99] 0.8464344 4.9042465