Function to generate random variates of the TSS distribution.

Generates n random numbers distributed according of the tempered stable subordinator distribution.

rTSS(
  n,
  alpha = NULL,
  delta = NULL,
  lambda = NULL,
  theta = NULL,
  methodR = "AR",
  k = 10000
)

Arguments

n

sample size (integer).

alpha

Stability parameter. A real number between 0 and 1.

delta

Scale parameter. A real number > 0.

lambda

Tempering parameter. A real number > 0.

theta

Parameters stacked as a vector.

methodR

A String. Either "AR" or "SR".

k

integer: the level of truncation, if methodR == "SR". 10000 by default.

Value

Generates n random numbers.

Details

theta denotes the parameter vector (alpha, delta, lambda). Either provide the parameters alpha, delta, lambda individually OR provide theta. "AR" stands for the Acceptance-Rejection Method and "SR" for a truncated infinite shot noise series representation. "AR" is the standard method used. For more details, see references.

References

Massing, Till (2022), 'Parametric Estimation of Tempered Stable Laws'

Kawai, R & Masuda, H (2011), 'On simulation of tempered stable random variates' doi:10.1016/j.cam.2010.12.014

Examples

rTSS(100,0.5,1,1)
#>   [1] 2.0440881 1.4048513 1.5300772 1.3472235 5.2672478 3.2682069 3.3924474
#>   [8] 1.6611119 1.4985139 2.6260454 3.8171909 1.7508057 0.8014868 0.5263920
#>  [15] 1.1824585 3.0266135 0.9562988 4.1772325 1.0618441 0.6338435 1.0315593
#>  [22] 3.4085385 2.9798679 1.3859102 3.3001087 1.4818353 0.7590354 2.1672644
#>  [29] 3.5663824 0.8535206 1.5197280 2.9599402 1.8384767 1.7875643 2.1310538
#>  [36] 2.3192351 1.6742548 1.2179690 3.8759912 1.4086266 1.3144904 0.7815468
#>  [43] 2.6910916 1.1312862 0.7129231 1.7860778 1.0608409 1.4325348 2.2951458
#>  [50] 0.9118742 2.3368735 4.8651204 0.5409956 1.0312381 2.9568956 0.8195319
#>  [57] 1.0103045 2.1240292 1.3143002 0.6515854 1.1979026 0.7469572 0.6723574
#>  [64] 1.9456982 3.5852250 2.8475780 3.3600548 0.7735942 0.7246238 0.8518410
#>  [71] 1.6628901 0.8444647 2.4454616 4.0489772 1.3317212 0.9527499 1.4868076
#>  [78] 2.6201762 2.2630380 0.4304098 1.5567032 1.0532315 1.3982226 1.6029758
#>  [85] 1.5216678 1.6606989 2.4051413 0.9258804 1.5928330 1.9541470 1.5938463
#>  [92] 0.8317497 1.2005457 0.9689984 3.7625235 1.6193277 0.5698879 2.0638746
#>  [99] 0.4958584 5.0309054
rTSS(100,0.5,1,1,NULL,"SR",50)
#>   [1] 1.5882636 0.5770326 1.4359771 0.9223960 2.4797086 1.6556270 0.7480915
#>   [8] 3.1005647 1.0757831 1.4182305 3.9013617 1.8833659 3.2271891 1.7330165
#>  [15] 2.7184490 1.8282248 2.5660132 0.9067352 2.3580708 1.0096593 3.2396070
#>  [22] 0.8560186 0.9767776 1.1129997 3.2912288 1.2091010 3.1216446 1.7006901
#>  [29] 1.9148719 1.0230442 0.4980394 1.5578640 1.3619711 1.6317202 0.6651544
#>  [36] 1.0868815 1.9860812 2.1856367 1.0094577 4.1946365 1.4681968 1.0612940
#>  [43] 1.0857901 2.3870256 1.5656335 1.7107877 1.2400023 1.6001953 2.3601192
#>  [50] 0.9813069 1.7236236 1.5467401 1.6580780 1.9800091 2.5617352 1.1979767
#>  [57] 2.1149344 0.9268106 1.3304349 1.7472485 1.5685106 1.5123138 1.6696970
#>  [64] 0.6703485 1.0251159 3.3675866 1.7385158 1.8277068 0.4820970 2.6783585
#>  [71] 1.0161182 0.8913025 1.1003717 1.3935287 1.0070053 1.3432143 1.3006023
#>  [78] 1.2045877 2.6310972 1.5519658 1.2625262 1.0603260 1.7185527 0.8062451
#>  [85] 0.9983950 1.2619876 0.7789922 0.9479468 1.1380293 4.4199209 1.5577778
#>  [92] 1.1687893 1.4875449 0.7827541 4.2856672 0.4684847 2.5139144 1.4095146
#>  [99] 2.2955346 2.0765470