Function to generate random variates of the TSS distribution.

Generates n random numbers distributed according of the tempered stable subordinator distribution.

rTSS(
  n,
  alpha = NULL,
  delta = NULL,
  lambda = NULL,
  theta = NULL,
  methodR = "AR",
  k = 10000
)

Arguments

n

sample size (integer).

alpha

Stability parameter. A real number between 0 and 1.

delta

Scale parameter. A real number > 0.

lambda

Tempering parameter. A real number > 0.

theta

Parameters stacked as a vector.

methodR

A String. Either "AR" or "SR".

k

integer: the level of truncation, if methodR == "SR". 10000 by default.

Value

Generates n random numbers.

Details

theta denotes the parameter vector (alpha, delta, lambda). Either provide the parameters alpha, delta, lambda individually OR provide theta. "AR" stands for the Acceptance-Rejection Method and "SR" for a truncated infinite shot noise series representation. "AR" is the standard method used. For more details, see references.

References

Massing, Till (2022), 'Parametric Estimation of Tempered Stable Laws'

Kawai, R & Masuda, H (2011), 'On simulation of tempered stable random variates' doi:10.1016/j.cam.2010.12.014

Examples

rTSS(100,0.5,1,1)
#>   [1] 2.0363673 5.1150451 0.9563463 0.7644210 2.3547096 1.2814746 4.1696949
#>   [8] 2.3136563 2.2649104 1.6972174 1.1259181 1.5701278 1.5740003 1.3983354
#>  [15] 1.0225786 1.1117760 2.4072141 1.2390590 2.4351540 1.5461451 1.4502372
#>  [22] 2.7539457 2.0711859 1.4026413 1.0567657 3.4874146 1.2254821 2.8137614
#>  [29] 2.4697923 1.9585292 2.2615981 2.6006709 1.8975043 2.2044178 1.7850773
#>  [36] 0.8675857 2.1481535 1.5546694 1.8000522 2.9555562 0.9762566 2.8580272
#>  [43] 1.6680197 0.7994765 2.3786442 2.2623827 1.4499651 2.0993855 3.5819988
#>  [50] 0.6807093 0.8971232 1.6104360 1.1237487 1.0274414 1.9794621 0.6079061
#>  [57] 1.5738203 0.8233529 2.1743005 1.0497219 1.6117037 2.1333783 1.5276908
#>  [64] 1.7954776 1.9287823 2.1122639 2.3643555 1.0540583 0.8250562 2.4199151
#>  [71] 1.4293052 0.7431755 1.9479194 0.8221491 0.8389079 0.6657993 4.9931133
#>  [78] 3.5351976 0.6213913 1.8099990 2.5592076 0.6936954 2.2739916 1.4780423
#>  [85] 0.9511469 1.6466845 2.1239573 1.1268928 1.2189754 1.6686868 1.1557531
#>  [92] 1.8440102 2.1356447 1.8877571 1.3631626 3.2586187 2.2830186 2.2367304
#>  [99] 1.3939709 2.1273873
rTSS(100,0.5,1,1,NULL,"SR",50)
#>   [1] 1.5760045 2.0488242 1.8461400 2.8482886 0.6874343 0.7198727 2.6865657
#>   [8] 5.2894118 2.6741036 1.3277959 1.7445068 1.7600793 3.8596979 0.9388203
#>  [15] 1.7328617 1.0455761 0.5066287 1.5316196 0.7159568 2.0506411 2.5329525
#>  [22] 1.9162355 0.5690918 4.6601087 1.4423629 2.0857676 0.8244685 2.2452902
#>  [29] 0.8975596 2.1493600 1.4119845 1.0380866 2.2066825 1.3537179 0.5637728
#>  [36] 1.9937824 2.9014240 0.2568950 1.8809076 1.0053037 2.0019854 1.2011497
#>  [43] 2.2485284 0.6063934 1.1058497 1.1993255 0.7818216 0.6685430 1.9835122
#>  [50] 0.3147637 3.1692768 0.4949140 2.1862952 0.9618522 1.8304669 1.3962471
#>  [57] 0.9763217 1.1947846 1.9697476 3.0372623 2.1260579 3.6392898 0.8849197
#>  [64] 1.3894549 1.8822885 1.6463445 0.9161112 0.7950216 2.3101672 2.8059921
#>  [71] 1.8014170 1.6970276 1.6476899 2.0069430 0.9022687 2.5822544 3.3116228
#>  [78] 1.8114276 1.5068913 0.6165315 2.2271372 1.6232266 2.2136696 1.9069805
#>  [85] 1.5036046 1.8441288 1.4710790 1.6998778 2.2677377 0.7994297 0.4394420
#>  [92] 1.7926713 1.3654756 1.1231483 2.5756848 0.4332075 1.5739913 0.8464344
#>  [99] 4.9042465 1.1091040