Function to generate random variates of the TSS distribution.

Generates n random numbers distributed according of the tempered stable subordinator distribution.

rTSS(
  n,
  alpha = NULL,
  delta = NULL,
  lambda = NULL,
  theta = NULL,
  methodR = "AR",
  k = 10000
)

Arguments

n

sample size (integer).

alpha

Stability parameter. A real number between 0 and 1.

delta

Scale parameter. A real number > 0.

lambda

Tempering parameter. A real number > 0.

theta

Parameters stacked as a vector.

methodR

A String. Either "AR" or "SR".

k

integer: the level of truncation, if methodR == "SR". 10000 by default.

Value

Generates n random numbers.

Details

theta denotes the parameter vector (alpha, delta, lambda). Either provide the parameters alpha, delta, lambda individually OR provide theta. "AR" stands for the Acceptance-Rejection Method and "SR" for a truncated infinite shot noise series representation. "AR" is the standard method used. For more details, see references.

References

Massing, T. (2023), 'Parametric Estimation of Tempered Stable Laws'

Kawai, R & Masuda, H (2011), 'On simulation of tempered stable random variates' doi:10.1016/j.cam.2010.12.014

Examples

rTSS(100,0.5,1,1)
#>   [1] 2.6634063 1.1500578 2.2710713 3.0511958 1.6018036 1.0530591 1.1600047
#>   [8] 1.5051593 1.7776393 1.3093383 2.6348098 1.3783367 1.2363693 1.9916418
#>  [15] 2.8633906 2.1567512 1.3411283 1.2010941 1.3194534 2.3158209 0.8511028
#>  [22] 5.8553705 1.9444518 2.9545249 0.7049878 1.9451381 1.8793000 1.0358444
#>  [29] 2.4196174 2.2844409 2.2528863 1.3746438 2.2603328 1.0015227 1.5888496
#>  [36] 0.6480019 0.5092060 2.0462480 2.2978809 2.7168312 2.6592355 2.2678267
#>  [43] 2.9838868 2.5554539 1.0619350 0.7172605 2.0617777 0.8744616 1.6845095
#>  [50] 1.2110665 1.5355688 2.3732172 0.8070866 5.2937294 1.0362611 0.9831290
#>  [57] 1.4671145 2.8053730 1.4983258 0.8270671 1.0050990 2.0511495 2.2397547
#>  [64] 2.6176397 1.7407501 2.7456326 2.2556459 0.7966161 1.4705738 1.3281374
#>  [71] 2.5881218 2.4008468 2.0613209 0.9466398 3.5303178 4.6029800 2.3036219
#>  [78] 3.0705530 2.4988662 3.7983237 2.4838983 1.3331136 1.7166408 0.8727304
#>  [85] 0.9104272 2.2924041 1.8373003 1.3761406 2.1817094 1.6469192 2.9766981
#>  [92] 1.4102546 3.0329564 1.3109017 0.9908380 0.5344541 2.4890788 1.2408047
#>  [99] 1.4356759 1.0010446
rTSS(100,0.5,1,1,NULL,"SR",50)
#>   [1] 0.6860700 2.0411851 2.5083037 3.1764710 0.5315949 1.8526555 1.9658741
#>   [8] 2.1860727 0.6948526 1.9481239 1.2467767 1.4774703 1.1880343 0.3929350
#>  [15] 1.3875205 1.2764257 3.9359985 6.1745616 1.3222896 0.9870662 4.9079658
#>  [22] 1.6796296 1.4662822 0.6830533 0.6853492 2.8497687 1.2719810 1.5248144
#>  [29] 0.8254804 0.8455653 3.9785301 3.1835158 1.1670202 0.9046895 2.3214144
#>  [36] 1.8684486 1.0985139 3.0203760 0.8068846 1.0731150 1.5396671 1.2275208
#>  [43] 1.3808135 1.6621600 0.8840397 1.9112077 1.4278730 2.1077734 0.6362299
#>  [50] 6.7545703 0.9430706 0.4680446 1.2994037 2.3069491 1.5306653 0.8969922
#>  [57] 1.3291934 1.0579247 2.2975973 0.9398703 1.7870328 2.4379489 1.9585180
#>  [64] 2.5633987 1.4163667 2.0524929 0.9403576 1.1881591 2.1244852 1.0711328
#>  [71] 1.5766262 2.9062319 0.8129022 1.6662762 1.6606269 1.0482336 2.0869778
#>  [78] 2.2693019 1.8168792 2.2423117 2.5596392 4.5306072 4.0758431 1.7826382
#>  [85] 2.8662696 1.7466043 1.3748772 0.9214059 1.7706536 0.9406344 0.9328480
#>  [92] 2.3194591 1.5934814 0.9171205 2.2127847 0.6624335 2.0436412 1.4101097
#>  [99] 2.3847400 1.4157262