Function to generate random variates of the TSS distribution.

Generates n random numbers distributed according of the tempered stable subordinator distribution.

rTSS(
  n,
  alpha = NULL,
  delta = NULL,
  lambda = NULL,
  theta = NULL,
  methodR = "AR",
  k = 10000
)

Arguments

n

sample size (integer).

alpha

Stability parameter. A real number between 0 and 1.

delta

Scale parameter. A real number > 0.

lambda

Tempering parameter. A real number > 0.

theta

Parameters stacked as a vector.

methodR

A String. Either "AR" or "SR".

k

integer: the level of truncation, if methodR == "SR". 10000 by default.

Value

Generates n random numbers.

Details

theta denotes the parameter vector (alpha, delta, lambda). Either provide the parameters alpha, delta, lambda individually OR provide theta. "AR" stands for the Acceptance-Rejection Method and "SR" for a truncated infinite shot noise series representation. "AR" is the standard method used. For more details, see references.

References

Massing, T. (2023), 'Parametric Estimation of Tempered Stable Laws'

Kawai, R & Masuda, H (2011), 'On simulation of tempered stable random variates' doi:10.1016/j.cam.2010.12.014

Examples

rTSS(100,0.5,1,1)
#>   [1] 1.6323375 1.3657765 1.3741514 1.2227996 1.3465096 1.2838966 3.3112045
#>   [8] 0.8756158 2.9394987 1.1514772 1.5196888 2.0953777 2.1904605 3.1385368
#>  [15] 1.2494331 1.8208203 0.8997275 1.0083072 0.6908096 2.7667694 4.2208238
#>  [22] 1.0849140 1.4726869 2.5995910 2.1797970 1.2732512 1.2950217 4.2934417
#>  [29] 2.0221813 3.2534833 3.6163262 3.1895766 1.6234495 1.3054140 1.1918688
#>  [36] 0.9990983 1.4651918 1.8281107 2.7110874 0.8420592 2.7979917 2.7891148
#>  [43] 1.5923369 0.7069871 1.8114035 1.2486695 1.1281430 1.0035351 0.7614477
#>  [50] 1.3307511 2.2895492 2.5972117 0.6709428 1.3416112 1.6154668 0.5231238
#>  [57] 2.8053407 0.9654455 1.7162297 1.2398392 1.3808506 1.8596199 0.4692675
#>  [64] 0.7318040 2.8962624 0.8926677 0.6714150 1.7040155 1.2331007 0.5096005
#>  [71] 1.5372797 1.3098870 1.5969980 1.1956747 2.0591018 1.0129906 0.8714762
#>  [78] 1.6554808 2.5370728 0.9579535 3.0256704 2.4028241 1.9828426 2.0210412
#>  [85] 1.5555061 3.7343834 4.5164462 0.9396777 0.7468095 1.6011380 2.2031832
#>  [92] 1.1496941 1.2446190 1.3369120 1.6287030 4.2347754 3.8266477 0.6979203
#>  [99] 0.9661528 1.7917934
rTSS(100,0.5,1,1,NULL,"SR",50)
#>   [1] 1.3236320 1.3048037 2.1447813 0.6482600 1.6651216 2.4083878 2.8412871
#>   [8] 0.8562726 1.6700963 0.8443187 0.8020044 1.3336682 2.8107314 1.5770181
#>  [15] 1.0642019 0.6082546 0.7769458 0.8887385 1.1578725 3.8291804 1.0295676
#>  [22] 1.7056403 2.9886254 0.9118317 4.0505410 1.2240291 2.2316639 1.4950632
#>  [29] 1.2659392 3.3347833 1.1193089 3.6525495 5.6081594 0.9160321 0.7780987
#>  [36] 3.2566262 1.1924683 3.4626048 3.1984415 2.1582807 3.4992753 2.8789359
#>  [43] 1.6229291 2.0416731 2.8131133 1.6182803 2.4521452 2.0851543 1.1300764
#>  [50] 1.7181386 2.4129754 1.3703138 0.8039372 1.6892775 1.7705639 1.5236019
#>  [57] 2.1074078 1.5553566 0.8501047 2.3527323 0.8752620 1.4009336 1.2408172
#>  [64] 1.8638096 0.8549938 2.3830044 1.3818514 1.0165542 3.2644516 0.7595065
#>  [71] 1.3326324 0.9264891 2.0447706 1.2664364 1.3196011 2.0355480 5.3254430
#>  [78] 0.9523190 0.9941440 1.9176517 1.6557198 2.5885717 3.0479015 2.2544515
#>  [85] 2.4650600 1.4243663 0.3755334 0.4788290 0.4847847 3.0587743 0.5871691
#>  [92] 2.7279908 2.2856408 5.6593852 1.2033259 1.6605924 1.7826664 1.5081666
#>  [99] 3.0249609 1.8049881