Function to generate random variates of the TSS distribution.

Generates n random numbers distributed according of the tempered stable subordinator distribution.

rTSS(
  n,
  alpha = NULL,
  delta = NULL,
  lambda = NULL,
  theta = NULL,
  methodR = "AR",
  k = 10000
)

Arguments

n

sample size (integer).

alpha

Stability parameter. A real number between 0 and 1.

delta

Scale parameter. A real number > 0.

lambda

Tempering parameter. A real number > 0.

theta

Parameters stacked as a vector.

methodR

A String. Either "AR" or "SR".

k

integer: the level of truncation, if methodR == "SR". 10000 by default.

Value

Generates n random numbers.

Details

theta denotes the parameter vector (alpha, delta, lambda). Either provide the parameters alpha, delta, lambda individually OR provide theta. "AR" stands for the Acceptance-Rejection Method and "SR" for a truncated infinite shot noise series representation. "AR" is the standard method used. For more details, see references.

References

Massing, Till (2022), 'Parametric Estimation of Tempered Stable Laws'

Kawai, R & Masuda, H (2011), 'On simulation of tempered stable random variates' doi:10.1016/j.cam.2010.12.014

Examples

rTSS(100,0.5,1,1)
#>   [1] 4.2511483 1.3703861 2.5973093 2.0139614 2.7120122 2.3712265 1.1711241
#>   [8] 2.0792050 1.4936013 0.9995226 2.3474584 2.8602498 2.1238128 1.3360797
#>  [15] 1.3822698 1.0496670 0.9763668 0.9760709 2.7475257 0.9332752 1.2848605
#>  [22] 2.5014686 2.3575114 3.0127467 1.0216600 1.2330281 2.4279217 0.4080413
#>  [29] 1.4336213 2.2814476 2.8747605 3.5712452 0.7407802 1.6743773 0.9825771
#>  [36] 0.8912080 0.9344805 1.4978696 2.0807129 1.5618291 1.7302371 2.5544773
#>  [43] 0.9125285 1.9894427 2.1466106 1.0047669 1.9722637 2.7834281 1.2081251
#>  [50] 2.9533575 0.9436449 1.0445643 0.6714374 2.7165811 0.9574357 1.8147018
#>  [57] 2.4018090 1.9685533 2.5534299 2.4958833 4.6431031 1.6750565 2.9847357
#>  [64] 1.1907655 1.4642408 2.8697968 1.6276921 2.8881463 1.4601082 3.1631668
#>  [71] 1.1264391 1.1191123 3.8602989 3.0984543 2.0512277 1.5596906 1.6246904
#>  [78] 2.8948366 2.7806148 0.6910053 1.9136909 0.6836110 1.5316963 1.3409029
#>  [85] 1.3831369 1.9609408 1.2926856 1.5768171 0.7064508 1.1709344 1.3864818
#>  [92] 1.4846180 1.5588484 1.5892313 0.6021401 1.7120175 1.4520707 1.6292044
#>  [99] 1.5049991 2.0473233
rTSS(100,0.5,1,1,NULL,"SR",50)
#>   [1] 1.0071224 3.9391368 2.9260678 1.6438166 0.8608394 2.1276672 1.7197615
#>   [8] 2.4629888 1.1804032 0.9708607 1.0895644 0.5384941 1.6269501 0.6989806
#>  [15] 6.0676609 1.0558760 3.0686314 1.3900018 1.2804002 2.4976001 0.6339115
#>  [22] 1.3798641 1.8195780 1.3556717 1.4008299 2.9498649 1.8814266 1.5688128
#>  [29] 1.0801718 0.9563738 0.6429376 6.1700441 0.9983924 1.0058565 1.8058955
#>  [36] 1.2956635 1.7708822 0.5356847 0.9192219 0.4220636 2.7860685 1.0387145
#>  [43] 1.0401677 1.4314779 1.0659372 1.1681280 1.6034732 1.4790633 1.9549907
#>  [50] 2.4948722 3.2253406 1.6312083 1.3245528 1.6654963 1.3056686 0.6955957
#>  [57] 1.2762274 1.0825518 1.6377657 3.1961139 1.7451402 1.3925491 1.3126619
#>  [64] 2.9536104 1.4067068 1.4165621 1.2227719 1.2431216 1.9669819 0.8277159
#>  [71] 0.8391543 0.4461831 1.5241930 1.2396973 1.1595880 1.3399942 1.1586417
#>  [78] 0.9203468 1.2214606 0.7730007 0.6640765 0.7525995 1.1303922 2.0243836
#>  [85] 1.5316465 1.7676763 1.1317491 1.0772977 1.0494996 2.0445439 1.4054946
#>  [92] 1.8481565 3.2920858 1.6240419 0.7086154 1.3340435 3.3931015 5.2466964
#>  [99] 1.1937383 3.4890972