Function to generate random variates of the TSS distribution.

Generates n random numbers distributed according of the tempered stable subordinator distribution.

rTSS(
  n,
  alpha = NULL,
  delta = NULL,
  lambda = NULL,
  theta = NULL,
  method = "AR",
  k = 10000
)

Arguments

n

sample size (integer).

alpha

Stability parameter. A real number between 0 and 1.

delta

Scale parameter. A real number > 0.

lambda

Tempering parameter. A real number > 0.

theta

Parameters stacked as a vector.

method

A String. Either "AR" or "SR".

k

integer: the level of truncation, if method == "SR". 10000 by default.

Value

Generates n random numbers.

Details

theta denotes the parameter vector (alpha, delta, lambda). Either provide the parameters alpha, delta, lambda individually OR provide theta. "AR" stands for the Acceptance-Rejection Method and "SR" for a truncated infinite shot noise series representation. "AR" is the standard method used. For more details, see references.

References

Massing, Till (2022), 'Parametric Estimation of Tempered Stable Laws'

Kawai, R & Masuda, H (2011), 'On simulation of tempered stable random variates' doi:10.1016/j.cam.2010.12.014

Examples

rTSS(100,0.5,1,1)
#>   [1] 1.8098335 1.7456654 0.9851650 3.4152722 1.1078734 0.8789195 1.6568377
#>   [8] 1.5258713 1.6729084 1.4936860 1.0868637 3.6255495 2.4024515 1.4088672
#>  [15] 1.8962104 2.2434816 1.9203924 1.5128202 0.7970817 1.0128738 0.4259809
#>  [22] 2.1800696 3.4384610 2.6021649 1.6721887 1.0855649 0.8059918 1.0656934
#>  [29] 1.1156306 0.7241432 0.6479879 1.0911596 1.0239076 1.1522545 0.8414924
#>  [36] 1.1306866 1.7298121 2.7937940 1.3785883 1.5162381 3.2882642 1.1738800
#>  [43] 3.0281149 5.5781730 1.0905655 0.9699351 0.7291611 0.9808815 1.2466741
#>  [50] 1.6142413 1.7463204 1.7458455 1.1134241 1.8382268 1.3739105 1.3693058
#>  [57] 1.5083588 0.7918593 1.6356908 1.1648094 0.7245895 0.4609151 1.5401162
#>  [64] 1.9190627 1.4484543 1.4960687 1.6066605 1.5882052 1.7832864 1.7457059
#>  [71] 1.2216184 1.9141492 1.7654496 1.1825254 1.1335643 1.4220606 1.8234726
#>  [78] 1.7783750 1.4040496 1.4194639 1.6897595 2.1865246 2.6499643 2.3274044
#>  [85] 4.0149023 2.9121311 1.0984744 1.5254369 1.2495770 1.6893935 0.5879279
#>  [92] 1.1007342 1.2742012 1.0843209 1.5356944 3.1439477 1.1540613 3.1963083
#>  [99] 2.8325835 0.9517232
rTSS(100,0.5,1,1,NULL,"SR",50)
#>   [1] 2.2849407 2.0876068 0.8236743 0.5434041 1.1871554 2.6086143 0.5719311
#>   [8] 1.5274266 1.9387790 3.1539331 0.6180649 0.9889605 2.0516548 0.4798287
#>  [15] 1.5255955 1.5444380 1.4357854 1.3400382 2.9057947 1.9053170 1.4137021
#>  [22] 5.0179731 1.5066398 1.8532291 1.4319005 1.1372614 4.1449675 0.9452749
#>  [29] 1.2930729 2.5649493 2.7364856 1.1481699 1.9334296 0.8118890 1.1927341
#>  [36] 0.9011715 1.7978026 1.8812171 0.9685319 2.0871169 2.2781511 1.1004875
#>  [43] 3.3398714 0.7370483 2.8686393 0.9457920 1.3624998 1.3423709 0.7985808
#>  [50] 0.9289411 3.4012350 1.4534930 2.7081960 1.3272092 2.1115735 3.6635781
#>  [57] 0.5499009 1.1574355 2.3258019 1.2451012 2.2979525 0.9485099 2.3496692
#>  [64] 1.6486623 2.3203436 2.1966360 2.0139086 1.1088246 7.2679155 0.8293659
#>  [71] 2.1724247 2.6824054 0.7174120 1.9108149 0.9993661 2.1340490 1.3372122
#>  [78] 3.8941703 1.8543796 1.2518978 1.5722015 1.7674357 1.1930672 1.1518705
#>  [85] 1.2421500 1.4780158 1.1695226 0.9727027 1.5536399 1.5573951 0.9703770
#>  [92] 2.4922379 1.2741068 2.1151807 0.6836009 3.0208207 1.8101128 1.3010993
#>  [99] 1.3687127 0.6855620