Function to generate random variates of the TSS distribution.

Generates n random numbers distributed according of the tempered stable subordinator distribution.

rTSS(
  n,
  alpha = NULL,
  delta = NULL,
  lambda = NULL,
  theta = NULL,
  method = "AR",
  k = 10000
)

Arguments

n

sample size (integer).

alpha

Stability parameter. A real number between 0 and 1.

delta

Scale parameter. A real number > 0.

lambda

Tempering parameter. A real number > 0.

theta

Parameters stacked as a vector.

method

A String. Either "AR" or "SR".

k

integer: the level of truncation, if method == "SR". 10000 by default.

Value

Generates n random numbers.

Details

theta denotes the parameter vector (alpha, delta, lambda). Either provide the parameters alpha, delta, lambda individually OR provide theta. "AR" stands for the Acceptance-Rejection Method and "SR" for a truncated infinite shot noise series representation. "AR" is the standard method used. For more details, see references.

References

Massing, Till (2022), 'Parametric Estimation of Tempered Stable Laws'

Kawai, R & Masuda, H (2011), 'On simulation of tempered stable random variates' doi:10.1016/j.cam.2010.12.014

Examples

rTSS(100,0.5,1,1)
#>   [1] 1.5302088 2.6056943 0.8301812 1.5853036 3.4020800 0.7600959 2.9988932
#>   [8] 0.9385657 0.6277222 2.0333299 1.3016222 5.2774787 0.4148700 1.8705548
#>  [15] 1.1640652 3.1729450 0.7453909 1.6568762 0.7971433 2.8276063 1.5566001
#>  [22] 4.6145251 2.3893250 1.2918153 0.7836403 3.3570780 1.1391125 1.9216217
#>  [29] 2.3810154 1.7117929 2.1225409 0.9999421 2.1057885 1.3309929 3.7576552
#>  [36] 1.8000374 0.9377894 1.2969732 1.5709549 2.2486015 1.5100974 0.8161351
#>  [43] 1.1850592 0.9548262 1.3041879 2.3247166 1.9307976 1.2208660 2.7954317
#>  [50] 2.4738778 4.0803611 4.9436973 1.3029090 1.9370419 0.9554760 1.4976961
#>  [57] 1.9983303 1.8746022 0.6828149 0.8024045 0.5338475 0.8989530 1.9150704
#>  [64] 2.5249530 3.0359132 1.0245741 3.3871340 2.6398427 3.1417410 0.7302338
#>  [71] 1.9748808 1.1205770 0.6896968 0.8724898 1.7967385 2.5552419 1.7287578
#>  [78] 1.2931782 3.2553473 1.0383058 1.3595574 0.8959709 1.5335014 1.6996133
#>  [85] 0.7635592 0.9996157 1.7522377 1.7919300 0.6125549 2.9473181 2.2672055
#>  [92] 3.1440616 1.2472971 1.1793720 1.9920625 2.4421837 2.8592111 2.1452799
#>  [99] 0.4356478 1.2342742
rTSS(100,0.5,1,1,NULL,"SR",50)
#>   [1] 0.9095651 2.3001867 2.0710465 1.8792494 0.5860826 0.5334098 0.4671657
#>   [8] 1.2055169 1.5387645 0.7547562 0.5590505 0.3621906 4.1377265 0.5872289
#>  [15] 1.5846941 3.1803694 0.3775860 1.4490357 1.8766498 0.8524866 0.9646668
#>  [22] 1.5743491 1.4705447 1.5984301 4.3295017 1.4186270 1.0964910 2.3753624
#>  [29] 1.7463966 1.1315004 0.8552727 1.9488865 0.9658620 1.1086204 0.2869598
#>  [36] 1.6738693 1.6439705 0.9881550 1.5152334 2.4891115 3.0546619 1.3321195
#>  [43] 2.0349713 1.9092551 3.0454268 2.0343527 0.7551732 2.1010872 1.9380273
#>  [50] 1.8361341 2.5394785 1.1794782 1.6659454 1.5965275 1.1423645 1.4516951
#>  [57] 1.6436313 1.7777529 0.7112195 1.3233566 1.5794334 1.8716398 0.3163335
#>  [64] 2.3202883 1.0032386 0.8846060 1.7666055 2.6424123 0.8099041 1.9079420
#>  [71] 5.2048839 1.1538678 2.6973345 0.7474706 1.1138002 0.5435188 2.0994566
#>  [78] 1.8474247 1.7803925 0.4989104 2.6572612 1.7579524 0.8563664 1.7497798
#>  [85] 2.3063104 1.4105186 1.3989365 0.7271267 0.9605514 0.8745385 1.4471126
#>  [92] 2.4164103 1.0277340 0.6836528 2.3580049 2.2213373 1.5004029 0.5966492
#>  [99] 1.3537981 2.3637694