Function to generate random variates of the TSS distribution.

Generates n random numbers distributed according of the tempered stable subordinator distribution.

rTSS(
  n,
  alpha = NULL,
  delta = NULL,
  lambda = NULL,
  theta = NULL,
  method = "AR",
  k = 10000
)

Arguments

n

sample size (integer).

alpha

Stability parameter. A real number between 0 and 1.

delta

Scale parameter. A real number > 0.

lambda

Tempering parameter. A real number > 0.

theta

Parameters stacked as a vector.

method

A String. Either "AR" or "SR".

k

integer: the level of truncation, if method == "SR". 10000 by default.

Value

Generates n random numbers.

Details

theta denotes the parameter vector (alpha, delta, lambda). Either provide the parameters alpha, delta, lambda individually OR provide theta. "AR" stands for the Acceptance-Rejection Method and "SR" for a truncated infinite shot noise series representation. "AR" is the standard method used. For more details, see references.

References

Massing, Till (2022), 'Parametric Estimation of Tempered Stable Laws'

Kawai, R & Masuda, H (2011), 'On simulation of tempered stable random variates' doi:10.1016/j.cam.2010.12.014

Examples

rTSS(100,0.5,1,1)
#>   [1] 0.7674872 1.3251228 1.5011183 1.7542180 1.3406815 2.1705807 1.5098717
#>   [8] 2.5734378 1.5511887 2.7309537 1.0350581 1.2633839 0.6566517 1.8806165
#>  [15] 2.4303999 2.7471592 0.7707577 0.9960180 2.3364160 2.0462932 1.2887779
#>  [22] 1.0342739 1.3720934 2.6804174 2.7673433 2.9282486 1.3406192 1.3402843
#>  [29] 1.4365320 4.0783912 2.5025827 1.0766357 2.8183933 0.9550123 0.9800882
#>  [36] 0.5816445 1.1379278 1.5028419 1.8728362 1.6594030 0.8033085 2.3519859
#>  [43] 1.9870032 1.7313393 2.1124042 3.1956909 1.3492145 1.3437102 0.8941245
#>  [50] 2.1209109 1.6818960 4.7016194 1.0619268 5.3118856 3.4999192 1.2608821
#>  [57] 3.4371566 2.2254751 0.9299289 2.2629244 1.0245610 1.6389268 1.7746803
#>  [64] 1.2320861 1.9299789 0.9596407 1.0065697 2.5909440 2.4679468 1.4199776
#>  [71] 0.7386574 2.2433744 0.9123078 1.2314420 0.9673643 1.0999966 1.2821880
#>  [78] 1.6305211 1.2356471 0.7564688 1.5599709 1.0252814 1.1094023 1.2931874
#>  [85] 1.5578330 1.5931640 1.3579888 1.5072300 0.7359489 1.4918514 1.3625995
#>  [92] 1.5528022 1.9544686 1.0321650 1.0754197 2.0392129 1.1162524 1.4185352
#>  [99] 1.6287621 0.9446775
rTSS(100,0.5,1,1,NULL,"SR",50)
#>   [1] 0.6635260 1.0662494 1.0130157 0.8604597 0.5559365 1.5350381 2.1947995
#>   [8] 1.3261421 2.0647707 0.6166164 0.4838294 0.8236060 2.6985768 0.7517042
#>  [15] 1.5325808 0.6415755 1.2557605 0.5506765 1.7793886 1.2847610 0.7465915
#>  [22] 1.2352913 1.7921445 0.9665543 2.8418259 2.3698851 1.2488236 1.9492700
#>  [29] 2.2161932 0.9244260 1.3986817 2.3958091 2.8519166 1.8434752 2.9350004
#>  [36] 1.2931039 1.1938057 0.5957707 2.0524680 4.1890477 2.4024558 0.4780967
#>  [43] 1.6178683 1.9837803 0.8711781 1.1177561 2.2951229 1.1120574 0.9181896
#>  [50] 1.3373220 2.1443747 3.1985375 0.8414758 1.1015466 0.8606900 0.6381497
#>  [57] 0.8952209 1.3833994 1.4214658 3.4611069 0.5777841 0.9263077 0.9656979
#>  [64] 1.9369214 0.9882053 1.3635467 1.2124307 1.6871435 1.5056073 1.4517842
#>  [71] 1.9403599 1.5953633 1.4932202 0.9434153 1.8311343 0.8356066 0.7099901
#>  [78] 1.5387710 0.7541728 0.9837384 1.6460116 1.2679478 2.0182940 0.5418398
#>  [85] 0.9192219 0.4220636 2.7860685 1.0387145 1.0401677 1.4314779 1.0659372
#>  [92] 1.1681280 1.6034732 1.4790633 1.9549907 2.4948722 3.2253406 1.6312083
#>  [99] 1.3245528 1.6654963