Function to generate random variates of the TSS distribution.

Generates n random numbers distributed according of the tempered stable subordinator distribution.

rTSS(
  n,
  alpha = NULL,
  delta = NULL,
  lambda = NULL,
  theta = NULL,
  methodR = "TM",
  k = 10000
)

Arguments

n

sample size (integer).

alpha

Stability parameter. A real number between 0 and 1.

delta

Scale parameter. A real number > 0.

lambda

Tempering parameter. A real number > 0.

theta

Parameters stacked as a vector.

methodR

A String. Either "TM", "AR" or "SR".

k

integer: the level of truncation, if methodR == "SR". 10000 by default.

Value

Generates n random numbers.

Details

theta denotes the parameter vector (alpha, delta, lambda). Either provide the parameters alpha, delta, lambda individually OR provide theta. "AR" stands for the Acceptance-Rejection Method and "SR" for a truncated infinite shot noise series representation. "TM" stands for Two Methods as two different methods are used depending on which will be faster. In this method the function copula::retstable() is called. "TM" is the standard method used.

It is recommended to check the generated random numbers once for each distribution using the density function. If the random numbers are shifted, e.g. for the method "SR", it may be worthwhile to increase k.

For more details, see references.

References

Massing, T. (2023), 'Parametric Estimation of Tempered Stable Laws'

Kawai, R & Masuda, H (2011), 'On simulation of tempered stable random variates' doi:10.1016/j.cam.2010.12.014

Hofert, M (2011), 'Sampling Exponentially Tilted Stable Distributions' doi:10.1145/2043635.2043638

See also

Examples

rTSS(100,0.5,1,1)
#>   [1] 1.4583757 1.6632501 2.0656764 1.2373118 2.0895579 1.0724561 1.0713649
#>   [8] 0.7705848 1.2440701 1.2822358 0.8628020 1.1125700 3.8362038 1.8056464
#>  [15] 1.1141210 3.7055805 1.3245349 1.0200101 2.6029855 2.9474003 0.7365623
#>  [22] 0.8381082 0.9336746 1.8346164 1.0044904 2.2608675 1.7072663 2.1712484
#>  [29] 4.8933917 4.2509487 1.6992963 1.4520361 3.7884889 3.0800224 2.3211423
#>  [36] 1.9168192 2.8215049 1.3664914 0.9683763 0.6347859 2.2636707 1.2590925
#>  [43] 1.7466283 0.5657153 0.8876130 1.7299324 0.3915999 1.7924967 1.2866924
#>  [50] 0.8852168 1.1917650 1.1237519 3.0993917 1.3652135 1.9167521 1.5136316
#>  [57] 0.8880191 1.4627200 1.5353104 0.5279576 1.2564876 2.0667096 0.8930116
#>  [64] 1.4466728 1.9880731 0.9297006 1.5119925 0.8877712 0.8393094 0.7725380
#>  [71] 1.2062304 1.3367447 0.6022954 1.8495038 1.4659997 1.1359290 1.1061656
#>  [78] 4.1773032 1.2725010 1.2904398 1.3451513 1.3987436 2.2482605 2.4351131
#>  [85] 0.8353403 0.5176949 1.2182143 0.8437278 1.4045914 0.5175722 2.2379647
#>  [92] 2.2267734 2.4220678 0.8774707 1.5773493 0.9838558 1.4611709 2.8334821
#>  [99] 0.5986422 1.1726710
rTSS(100,0.5,1,1,NULL,"SR",50)
#>   [1] 1.9453127 1.8466299 0.6907699 1.6245756 1.5118690 1.1680851 0.9925529
#>   [8] 2.0904955 3.2061681 1.9826629 1.5355394 0.4155417 2.2161755 2.6538276
#>  [15] 1.8136229 1.6907369 2.4287636 1.6759213 0.5248442 1.6338075 2.2772707
#>  [22] 4.7125421 0.8870366 0.9620239 2.7521315 1.0373965 1.1336556 2.2620787
#>  [29] 2.5101990 1.4436114 0.9994393 0.8275206 0.8960774 0.9736340 1.1763805
#>  [36] 2.2005444 1.5755204 2.4520042 1.2474552 1.7743722 0.9955159 1.2625734
#>  [43] 3.2333399 0.5506186 2.1172775 4.1561386 1.8415682 2.5729962 1.4417597
#>  [50] 3.1962129 0.7979705 4.7250759 0.2830478 1.3781311 0.5969741 1.1140808
#>  [57] 2.7534116 1.4367373 2.1797909 1.1484198 1.1463197 0.5088579 2.4802867
#>  [64] 1.5916848 1.7183683 1.6507296 0.9122264 1.3395311 1.6298262 1.7580876
#>  [71] 2.1388841 2.5772009 1.7708364 0.7559716 3.4032357 1.0268233 1.4217619
#>  [78] 4.1241101 1.2779820 1.6376790 3.2283201 1.7455832 1.5318637 1.0109317
#>  [85] 0.8607448 1.1492461 1.1579701 2.2879786 3.4984554 2.1855866 1.7882808
#>  [92] 2.9665427 1.3066292 2.1092196 1.9680124 1.8912956 1.5609094 1.6412738
#>  [99] 1.7407151 0.9302302